### 2025-02-23 #7838 @editorialbot
[PRE REVIEW]: COSMOS: A numerical relativity code specialized for PBH formation
**Submitting author:** @cmyoo (<a href="http://orcid.org/0000-0002-9928-4757">Chul-Moon Yoo</a>)
**Repository:** https://github.com/cmyoo/cosmos
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
