### 2025-02-23 #7838 @warrickball
Hi @cmyoo — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, could you explain the relationship between [COSMOS](https://github.com/cmyoo/cosmos) and [COSMOS-S](https://github.com/cmyoo/cosmos)? It looks to me like the first is for 3 spatial dimensions and the latter for 1 spatial dimension under spherical symmetry but it isn't clear why there are two separate code bases. At a glance I think there are a few thousand lines of shared C++, so it seem more natural to me to have a single package/repository that builds both executables (or has a build system in which the user can select) and the user can the decide what to build/run. (I've seen this done in codes in other niches of astronomy in the past.)

As for the paper failing to build, I expect the error is that your affiliation should be `"1, 2"` instead of `[1, 2]`. (I'm also aware this is our fault at JOSS: there was an error in the documentation that was fixed last week.) If that doesn't fix things, I'll debug in more detail.
