### 2025-02-24 #7838 @cmyoo
> In the meantime, could you explain the relationship between [COSMOS](https://github.com/cmyoo/cosmos) and [COSMOS-S](https://github.com/cmyoo/cosmos)? It looks to me like the first is for 3 spatial dimensions and the latter for 1 spatial dimension under spherical symmetry but it isn't clear why there are two separate code bases. At a glance I think there are a few thousand lines of shared C++, so it seem more natural to me to have a single package/repository that builds both executables (or has a build system in which the user can select) and the user can the decide what to build/run. (I've seen this done in codes in other niches of astronomy in the past.)

Thank you for the suggestion.
Our concept is to provide minimal functions in COSMOS and COSMOS-S for those supposed calculations. Then there are two reasons why we provide two separate code bases. First, we expect that not a few users are only interested in spherically symmetric settings, and the 3+1 dimensional simulation code is redundant for them. Second, although these two code bases share a basic straight-forward but long evaluation of basic equations, many specialized functions cannot be shared. Therefore we wanted to avoid confusion due to the existence of functions that are not necessarily needed for some users.
