+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
   - name: Chul-Moon Yoo
     orcid: 0000-0002-9928-4757
     # equal-contrib: true
-    affiliation: [ 1, 2 ] # (Multiple affiliations must be quoted)
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
   - name: Albert Escrivà
     orcid: 0000-0001-5483-8034
     # equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
