+++ paper/paper.bib
@@ -69,7 +69,8 @@
  publisher = {World Scientific Publishing Co. Pte. Ltd},
  year = {2016},
  address = {Singapore},
- isbn = {978-981-4699-71-6}
+ isbn = {978-981-4699-71-6},
+ doi = {10.1142/9692}
  }
 @Book{gourgoulhon20123+1,
  author = {Gourgoulhon, Eric},
@@ -77,7 +78,8 @@
  publisher = {Springer},
  year = {2012},
  address = {Berlin New York},
- isbn = {978-3-642-24525-1}
+ isbn = {978-3-642-24525-1},
+ doi = {10.1007/978-3-642-24525-1}
  }
 @article{Yoo:2024lhp,
     author = "Yoo, Chul-Moon",
@@ -263,12 +265,15 @@ journal = {Classical and Quantum Gravity}
 }
 @article{Shimada:2024eec,
     author = "Shimada, Masaaki and Escriv\'a, Albert and Saito, Daiki and Uehara, Koichiro and Yoo, Chul-Moon",
-    title = "{Primordial Black Hole Formation from Type II Fluctuations with Primordial Non-Gaussianity}",
+    title = "{Primordial black hole formation from type II fluctuations with primordial non-Gaussianity}",
     eprint = "2411.07648",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
-    month = "11",
-    year = "2024"
+    doi = "10.1088/1475-7516/2025/02/018",
+    journal = "JCAP",
+    volume = "02",
+    pages = "018",
+    year = "2025"
 }
 @article{Alcubierre:1999ab,
     author = "Alcubierre, Miguel and Brandt, Steven and Bruegmann, Bernd and Holz, Daniel and Seidel, Edward and Takahashi, Ryoji and Thornburg, Jonathan",
