+++ README.md
@@ -1,6 +1,7 @@
 [![C++](https://custom-icon-badges.herokuapp.com/badge/C++-f34b7d.svg?logo=Cplusplus&logoColor=white)]()
 [![Github All Releases](https://img.shields.io/github/downloads/atom/atom/total.svg?style=flat)]()
 [![BSD-3-Clause](https://custom-icon-badges.herokuapp.com/badge/license-BSD%203%20Clause-8BB80A.svg?logo=law&logoColor=white)]()
+ [![status](https://joss.theoj.org/papers/f03e0df3a83ae294aedc8629dddd88e8/status.svg)](https://joss.theoj.org/papers/f03e0df3a83ae294aedc8629dddd88e8)
 
 # COSMOS
 <!-- 
