### 2025-02-24 #7838 @cmyoo
> As for the paper failing to build, I expect the error is that your affiliation should be "1, 2" instead of [1, 2]. (I'm also aware this is our fault at JOSS: there was an error in the documentation that was fixed last week.) If that doesn't fix things, I'll debug in more detail.

I updated the paper.
