### 2025-02-24 #7838 @warrickball
Thanks for the context. I'll go ahead and see if I can find some precedent or get some guidance on how JOSS will handle two repos in a single review. As a note to any editor who picks this up, note that there are two repos under review: COSMOS and COSMOS-S, and both should be reviewed.
