### 2025-03-10 #7838 @warrickball
Hi @cmyoo, your paper is still on the waiting list, I'm afraid, but note that there's another error in the paper source. Taishi Ikeda's ORCID should be **0**000-0002-9076-1027.

If you fix that, you can try generating the draft again by starting a message here with `@editorialbot generate pdf`.
