### 2025-04-11 #7838 @fruzsinaagocs
👋 @amelialdrew, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
Note that there are two repositories to be reviewed here: [COSMOS](https://github.com/cmyoo/cosmos) and [COSMOS-S](https://github.com/cmyoo/cosmos-s).
