### 2025-04-15 #7838 @fruzsinaagocs
👋 @Sbozzolo, would you be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can look for other reviewers.
Note that there are two repositories to be reviewed here: [COSMOS](https://github.com/cmyoo/cosmos) and [COSMOS-S](https://github.com/cmyoo/cosmos-s).
