### 2025-04-15 #7838 @Sbozzolo
@fruzsinaagocs My first impression is that the project is really not ready for review: I don't see any documentation/tests/contributors guidelines.

In order to best utilize and respect reviewers's time, I would strongly recommend not going forward with this review until the authors have added enough material to feel confident that their project will clear [the review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html).
