### 2025-04-15 #7838 @ShaunFell
@fruzsinaagocs Upon closer inspection, I agree with Gabriele. There are no tests and the documentation linked in the repo is extremely scarce.
