### 2025-04-15 #7838 @fruzsinaagocs
@Sbozzolo @ShaunFell Thank you for spotting this, and apologies for not having noticed that earlier. I'll ask the authors to bring the code more compliant with the reviewer criteria before we proceed with the review.

@cmyoo, Please make sure that, as per the [JOSS submission guide](https://joss.readthedocs.io/en/latest/submitting.html#submission-process),

> the software complies with the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). In particular, your software should be full-featured, well-documented, and contain procedures (such as automated tests) for checking correctness.

This includes API-level documentation and tests. I see that you have tutorial-style examples and some integration tests already, but it's not clear what the desired behavior of the code is, i.e. there should be assertions to make sure the code doesn't produce unphysical results.
