### 2025-04-23 #7838 @cmyoo
Hi @fruzsinaagocs,
Thank you for your further comments. I tried to generate API documents using AI code assist, and uploaded it to docs/html directory of COSMOS. We didn't check it in detail yet, but could you please check if the format is okay?
