+++ docs/sources/test.h
@@ -1,0 +1,1298 @@
+/**
+ * @file cosmos.h
+ * @brief Header file for the BSSN evolution classes of the COSMOS code.
+ *
+ * Defines the base class Fmv0 for grid management, variable storage,
+ * finite differencing, and basic evolution utilities. Also defines derived
+ * classes Fmv (for specific initial data and boundary conditions) and
+ * Fmv1 (for adaptive mesh refinement).
+ *
+ * @version 1.00
+ * @author Chulmoon Yoo
+ */
+
+ #ifndef _COSMOS_H_
+ #define _COSMOS_H_
+ // ... (includes and definitions) ...
+ 
+ using namespace std;
+ 
+ /**
+  * @class Fmv0
+  * @brief Base class for Finite Mesh Variables version 0.
+  *
+  * This class manages the computational grid, stores dynamical variables
+  * (BSSN, gauge, fluid, scalar fields), provides methods for finite
+  * differencing, basic time stepping support, constraint checking,
+  * and memory management. It forms the foundation for specific simulation
+  * setups and mesh refinement.
+  */
+ class Fmv0{
+ 
+ protected:
+     // ... (protected member variables - not typically part of public API docs) ...
+ 
+ public:
+     /**
+      * @brief Constructor for the Fmv0 class.
+      * @param tabs Number of buffer zones (ghost zones).
+      * @param jupper Upper index boundary in x for the physical domain.
+      * @param jlower Lower index boundary in x for the physical domain.
+      * @param kupper Upper index boundary in y for the physical domain.
+      * @param klower Lower index boundary in y for the physical domain.
+      * @param lupper Upper index boundary in z for the physical domain.
+      * @param llower Lower index boundary in z for the physical domain.
+      * @param xupper Upper coordinate boundary in x.
+      * @param xlower Lower coordinate boundary in x.
+      * @param yupper Upper coordinate boundary in y.
+      * @param ylower Lower coordinate boundary in y.
+      * @param zupper Upper coordinate boundary in z.
+      * @param zlower Lower coordinate boundary in z.
+      * @param am Amplitude for inhomogeneous grid mapping (if used).
+      * @param fld Boolean flag to enable fluid evolution.
+      * @param scl Boolean flag to enable scalar field evolution.
+      * @param cuev Boolean flag to enable curvature evaluation.
+      *
+      * Initializes grid parameters, allocates memory for all variables,
+      * coordinate arrays, flags, and temporary storage. Sets up basic
+      * parameters and constants.
+      */
+     Fmv0(int tabs,int jupper,int jlower,int kupper,int klower,int lupper,int llower,
+     double xupper,double xlower,double yupper,double ylower,double zupper,double zlower,double am,bool fld, bool scl, bool cuev);
+ 
+     /**
+      * @brief Virtual destructor for the Fmv0 class.
+      *
+      * Deallocates all dynamically allocated memory associated with the grid,
+      * variables, flags, and coordinate arrays.
+      */
+     virtual ~Fmv0();
+ 
+     //--------------------------------------
+     // GET functions (Accessors)
+     //--------------------------------------
+ 
+     /** @brief Check if fluid evolution is enabled. */
+     bool get_fluidevo() const;
+     /** @brief Check if horizon formation has been detected (flag). */
+     bool get_hform() const;
+     /** @brief Check if excision is enabled. */
+     bool get_exc() const;
+     /** @brief Check if mesh refinement is active (flag, relevant for derived classes). */
+     bool get_mrf() const;
+ 
+     /** @brief Get the maximum grid index in x (including buffer). */
+     int get_jmax() const;
[…]
