### 2025-04-25 #7838 @cmyoo
Hi @fruzsinaagocs,
I learned how to the html API documents public and now it's available from https://cmyoo.github.io/cosmos/
