+++ docs/ahf2d_d.h
@@ -1,159 +1,0 @@
-/**
- * @file ahf2d.h
- * @brief Header file for the 2D Apparent Horizon Finder (AHF) class.
- * @details Implements methods to locate apparent horizons on spatial slices
- *          using a 2D spherical coordinate grid (theta, phi) and an
- *          iterative elliptic solver.
- * @version 1.0 // Or your current version
- * @date 2023-10-27 // Or the relevant date
- * @author Chulmoon Yoo // Or the primary author(s)
- */
-
- #ifndef AHF2D_H
- #define AHF2D_H
- 
- #include "cosmos.h" // Include if it needs access to Fmv0/Fmv grid data/methods
- #include <stdio.h>
- #include <math.h>
- // Include other necessary headers
- 
- /**
-  * @class Ahf2d
-  * @brief Implements a 2D Apparent Horizon Finder.
-  * @details Solves the apparent horizon equation Theta = 0, where Theta is the
-  *          expansion of outgoing null geodesics, on a spherical coordinate grid.
-  *          Uses an iterative method involving a Poisson-like solver.
-  */
- class Ahf2d {
- public:
-     // --- Member Variables ---
-     /** @brief Number of grid points in theta direction (excluding buffers). */
-     int ntheta;
-     /** @brief Number of grid points in phi direction (excluding buffers). */
-     int nphi;
-     /** @brief Grid spacing in theta. */
-     double dth;
-     /** @brief Grid spacing in phi. */
-     double dphi;
-     /** @brief Initial guess for the horizon radius (coordinate radius). */
-     double hini;
-     /** @brief Acceleration parameter eta for the elliptic solver. Also distinguishes AH/CH. */
-     double var; // Corresponds to etaah in par_ahf.d
-     /** @brief Mixing factor for updating horizon radius guess during iteration. */
-     double fac; // Corresponds to facn in par_ahf.d
-     /** @brief Tolerance for the Poisson solver part. */
-     double err_p;
-     /** @brief Tolerance for the main AH equation solver part. */
-     double err_e;
-     /** @brief Maximum number of iterations allowed for the solver. */
-     int ahloop;
- 
-     // Pointers to AHF internal grid arrays (Example for horizon shape 'h')
-     /** @brief Horizon shape function h(theta, phi). */
-     double **h;
-     /** @brief Expansion Theta(theta, phi). */
-     double **ths;
-     // Add other internal arrays if needed (e.g., for solver residuals, coefficients)
- 
-     /** @brief Pointer to the Fmv0/Fmv object containing the spacetime grid data. */
-     Fmv0 *fm; // Use Fmv0* or Fmv* depending on what data is needed
- 
-     // --- Constructor & Destructor ---
-     /**
-      * @brief Constructor for the Ahf2d class.
-      * @param ntheta_in Number of theta grid points.
-      * @param nphi_in Number of phi grid points.
-      * @param hini_in Initial guess for horizon radius.
-      * @param var_in Solver acceleration parameter eta (etaah).
-      * @param fac_in Radius update mixing factor (facn).
-      * @param err_p_in Poisson solver tolerance.
-      * @param err_e_in AH equation solver tolerance.
-      * @param ahloop_in Maximum solver iterations.
-      * @param fm_in Pointer to the Fmv0/Fmv object with spacetime data.
-      */
-     Ahf2d(int ntheta_in, int nphi_in, double hini_in, double var_in, double fac_in,
-           double err_p_in, double err_e_in, int ahloop_in, Fmv0 *fm_in);
- 
-     /**
-      * @brief Destructor for the Ahf2d class. Frees allocated memory.
-      */
-     ~Ahf2d();
- 
-     // --- Core Methods ---
-     /**
[…]
