+++ docs/sources/cosmos_d.h
@@ -1,374 +1,0 @@
-/**
- * @file cosmos.h
- * @brief Header file defining the core BSSN evolution classes (Fmv0, Fmv, Fmv1) for the COSMOS code.
- * @details Contains base class Fmv0 for grid variables and evolution,
- *          derived class Fmv for specific boundary conditions/initial data,
- *          and derived class Fmv1 for Fixed Mesh Refinement (FMR).
- * @version 1.00 // Or your current version
- * @date 2023-10-27 // Or the relevant date
- * @author Chulmoon Yoo // Or the primary author(s)
- */
-
- #ifndef COSMOS_H // Ensure include guards are present
- #define COSMOS_H
- 
- #include <stdio.h>
- #include <stdlib.h>
- #include <math.h>
- #include <string.h>
- // Include other necessary headers like complex.h if used
- 
- // --- Constants and Macros (Example) ---
- /** @brief Gravitational constant G (often set to 1 in simulations). */
- #define G_CONST 1.0
- /** @brief Speed of light c (often set to 1 in simulations). */
- #define C_LIGHT 1.0
- // Add other constants/macros if present
- 
- // --- Enums (Example if you have them) ---
- /** @brief Enum defining possible boundary condition types. */
- typedef enum {
-     BOUNDARY_PERIODIC,
-     BOUNDARY_REFLECTING,
-     BOUNDARY_OUTFLOW
-     // Add other boundary types
- } BoundaryType;
- 
- 
- // --- Fmv0 Class ---
- /**
-  * @class Fmv0
-  * @brief Base class for Finite Mesh Variables version 0.
-  * @details Handles the fundamental grid structure, memory allocation,
-  *          coordinate mapping, basic BSSN evolution steps (RHS calculation),
-  *          and Kreiss-Oliger dissipation. Does not implement specific
-  *          boundary conditions or mesh refinement.
-  */
- class Fmv0 {
- public: // Or protected, depending on your design
-     // --- Member Variables (Document important public/protected ones) ---
-     /** @brief Number of buffer/ghost zones. */
-     int tab;
-     /** @brief Minimum physical grid index in x. */
-     int jmin;
-     /** @brief Maximum physical grid index in x. */
-     int jmax;
-     // ... (similarly for kmin, kmax, lmin, lmax)
-     /** @brief Minimum coordinate value in x. */
-     double xmin;
-     /** @brief Maximum coordinate value in x. */
-     double xmax;
-     // ... (similarly for ymin, ymax, zmin, zmax)
-     /** @brief Grid spacing in x (uniform logical grid). */
-     double dx;
-     // ... (similarly for dy, dz)
-     /** @brief CFL factor for time step calculation. */
-     double cfl;
-     /** @brief Gauge parameter eta for 1+log lapse. */
-     double etaa;
-     /** @brief Gauge parameter eta for Gamma-driver shift. */
-     double etab;
-     /** @brief Gauge parameter eta_b (or similar) for Gamma-driver shift. */
-     double etabb;
-     /** @brief Kreiss-Oliger dissipation coefficient epsilon. */
-     double KOep;
-     /** @brief Initial Hubble parameter. */
-     double Hb;
-     /** @brief Fluid equation of state parameter w (if fluid enabled). */
-     double fluidw;
-     /** @brief MUSCL reconstruction parameter kappa (if fluid enabled). */
-     double kap_MUSCL;
-     /** @brief Minmod limiter parameter b (if fluid enabled). */
-     double b_minmod;
-     /** @brief Grid size for excision region (if enabled). */
-     int exg;
-     /** @brief Maximum simulation time. */
-     double tmax;
-     /** @brief Grid stretching parameter for coordinate mapping. */
[…]

[…]
