### 2025-04-28 #7838 @fruzsinaagocs
@cmyoo That looks like a positive change. Please make sure that:

- the generated API documentation covers all functions/classes/etc.
- the API documentation and the Wiki pages are linked to from the README.md
- the requirement about having adequate [tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) ideally means that the user can easily verify that the software works correctly, without too much domain knowledge. Right now this is a little difficult: the user has to reproduce figures and it's unclear how much deviation from the results of the figures are acceptable. Can you come up with tests that cover all functionality of the code and the expected behavior is asserted (e.g. code gives you a message/logs about whether it ran correctly)?
