### 2025-04-30 0222f87
[Add] add check_xkl function in the main files of examples
