+++ source/adiabatic_spherical/cosmos.cpp
@@ -40,6 +40,8 @@
 
 using namespace std;
 
+void check_xkl(Fmv *fmv);
+
 void printpack(Fmv0 **fmv0,int ln,int pk,int pl,
 ofstream& filex,
 ofstream& filey,
@@ -740,7 +742,10 @@ int main(int argc,char* argv[])
 		//output all data end
 	}
 	//final print end
-		
+
+	//comparing out_xkl.dat with exp_xkl.dat 
+	check_xkl(fmv);
+	
 	//finalize start
 	delete ahf;
 
@@ -1359,4 +1364,161 @@ int *lbs									//grid number for fmr region on z-axis
 	for(int i=0;i<ln;i++)
 	fileall << "##fmrzgnum="<< lbs[i] << endl;
 
+}
+
+//function to check difference of out_xkl.dat from exp_xkl.dat
+void check_xkl(Fmv *fmv)
+{
+	ifstream expfile("exp_xkl.dat");				//open exp file
+	if(expfile.fail()){
+		cout << "exp_xkl.dat is not found." << endl;
+		abort();
+	}
+
+	ifstream outfile("out_xkl.dat");				//open out file
+	if(outfile.fail()){
+		cout << "out_xkl.dat is not found." << endl;
+		abort();
+	}
+
+	string bufexp,bufout;
+
+	cout << "header texts in exp_xkl.dat" << endl;
+	getline(expfile, bufexp);
+	cout << bufexp << endl;
+	getline(expfile, bufexp);
+	cout << bufexp << endl;
+	
+	cout << "header texts in out_xkl.dat" << endl;
+	getline(outfile, bufout);
+	cout << bufout << endl;
+	getline(outfile, bufout);
+	cout << bufout << endl;
+
+	double vdiff[31];
+	for(int n=0;n<31;n++)
+	vdiff[n]=0.;
+
+	for(int j=fmv->get_jmin();j<=fmv->get_jmax();j++)
+	{
+		double vexp[31];
+		double vout[31];
+
+		getline(outfile, bufout);
+		getline(expfile, bufexp);
+		
+		if(fmv->get_fluidevo() && fmv->get_scalarevo())
+		{
+			// sscanf(bufout.data(),"%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf",
+			// &alpout,&bxout,&byout,&bzout,&bbxout,&bbyout,&bbzout,&gxxout,&gyyout,&gzzout,
+			// &gxyout,&gxzout,&gyzout,&waout,&akxxout,&akyyout,&akzzout,&akxyout,&akxzout,&akyzout,
+			// &ekout,&zgxout,&zgyout,&zgzout,&Eneout,&pxout,&pyout,&pzout,&Dout,&phiiout,&Piout);
+
+			sscanf(bufexp.data(),"%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf",
+			&vexp[0],&vexp[1],&vexp[2],&vexp[3],&vexp[4],&vexp[5],&vexp[6],&vexp[7],&vexp[8],&vexp[9],
[…]
