+++ docs/sources/cosmos_d.cpp
@@ -1,0 +1,1539 @@
+/**
+ * @file d:\chulmoon\cwork\COSMOS\cosmos_v1.00\source\cosmos.cpp
+ * @brief Main program file for the COSMOS numerical relativity simulation code.
+ * @author Chulmoon Yoo
+ * @version 1.00
+ *
+ * @details
+ * This file serves as the entry point and main driver for the COSMOS simulation,
+ * which evolves spacetime using the BSSN formulation, potentially coupled with
+ * fluid or scalar fields. It operates on a Cartesian grid, potentially utilizing
+ * Fixed Mesh Refinement (FMR) and simulating a quarter region with reflection boundaries.
+ *
+ * Key responsibilities include:
+ * - Reading simulation parameters from input files (`par_ini.d`, `par_fmr.d`, `par_ahf.d`).
+ * - Initializing the grid hierarchy and field variables, either from scratch based
+ *   on parameters or by loading data from a continuation file (`out_all.dat`).
+ * - Executing the main time evolution loop using a 4th-order Runge-Kutta scheme.
+ * - Managing the addition and evolution of FMR layers based on specified criteria.
+ * - Periodically invoking the Apparent Horizon Finder (AHF) to locate horizons.
+ * - Handling data output, including 1D/2D slices, constraint violations,
+ *   AH properties, and full simulation state for restarts.
+ *
+ * @note This code utilizes OpenMP for parallelization. Ensure the `OMP_NUM_THREADS`
+ *       environment variable is set appropriately before execution.
+ *
+ * @see cosmos.h, ahf2d.h
+ */
+//**********************************************************************************************//
+//----------------------------------------------------------------------------------------------//
+//                     @@@@@   @@@@    @@@@    @@@ @@    @@@@    @@@@                           //
+//                   @@@      @@   @  @@   @  @@@ @@ @  @@   @  @@   @                          //
+//                   @@       @@   @   @@@    @@  @  @  @@   @   @@@                            //
+//                   @@       @@   @ @    @@  @@     @  @@   @ @    @@                          //
+//                     @@@@@   @@@@   @@@@@   @@     @   @@@@   @@@@@                           //
+//----------------------------------------------------------------------------------------------//
+//**********************************************************************************************//
+//----------------------------------------------------------------------------------------------//
+//                 cosmos: quarter region, mesh-refinement                                      //
+//                                             ver. 1.00           coded by Chulmoon Yoo        //
+//                                                                                              //
+//----------------------------------------------------------------------------------------------//
+// Main file   :: cosmos.cpp                                                                    //
+// header      :: cosmos.h ahf2d.h                                                              //
+// definition  :: cosmos_bssn.cpp cosmos_initial.cpp cosmos_output.cpp                          //
[…]
