+++ source/scalar_iso/cosmos.cpp
@@ -742,8 +742,13 @@ int main(int argc,char* argv[])
 		//output all data end
 	}
 	//final print end
-		
-	//comparing out_xkl.dat with exp_xkl.dat 
+
+	ofstream filecheck("out_check_xkl.dat");						//as functions of x on (k,l) for check
+	filecheck << "## nmax=" << nxmax << " xmax=" << xmax <<  endl;	
+	fmv->dyntoprim();
+	fmv->print_x(filecheck,pk,pl);
+	
+	//comparing out_check_xkl.dat with exp_xkl.dat 
 	check_xkl(fmv);
 
 	//finalize start
@@ -1376,9 +1381,9 @@ void check_xkl(Fmv *fmv)
 		abort();
 	}
 
-	ifstream outfile("out_xkl.dat");				//open out file
+	ifstream outfile("out_check_xkl.dat");				//open out file
 	if(outfile.fail()){
-		cout << "out_xkl.dat is not found." << endl;
+		cout << "out_check_xkl.dat is not found." << endl;
 		abort();
 	}
 
@@ -1390,20 +1395,20 @@ void check_xkl(Fmv *fmv)
 	getline(expfile, bufexp);
 	cout << bufexp << endl;
 	
-	cout << "header texts in out_xkl.dat" << endl;
+	cout << "header texts in out_check_xkl.dat" << endl;
 	getline(outfile, bufout);
 	cout << bufout << endl;
 	getline(outfile, bufout);
 	cout << bufout << endl;
 
-	double vdiff[31];
-	for(int n=0;n<31;n++)
+	double vdiff[50];
+	for(int n=0;n<50;n++)
 	vdiff[n]=0.;
 
 	for(int j=fmv->get_jmin();j<=fmv->get_jmax();j++)
 	{
-		double vexp[31];
-		double vout[31];
+		double vexp[50];
+		double vout[50];
 
 		getline(outfile, bufout);
 		getline(expfile, bufexp);
@@ -1415,51 +1420,62 @@ void check_xkl(Fmv *fmv)
 			// &gxyout,&gxzout,&gyzout,&waout,&akxxout,&akyyout,&akzzout,&akxyout,&akxzout,&akyzout,
 			// &ekout,&zgxout,&zgyout,&zgzout,&Eneout,&pxout,&pyout,&pzout,&Dout,&phiiout,&Piout);
 
-			sscanf(bufexp.data(),"%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf",
+			sscanf(bufexp.data(),"%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf%lf",
 			&vexp[0],&vexp[1],&vexp[2],&vexp[3],&vexp[4],&vexp[5],&vexp[6],&vexp[7],&vexp[8],&vexp[9],
 			&vexp[10],&vexp[11],&vexp[12],&vexp[13],&vexp[14],&vexp[15],&vexp[16],&vexp[17],&vexp[18],&vexp[19],
[…]

[…]
