### 2025-05-02 #7838 @cmyoo
@fruzsinaagocs,
Thank you for your continued support.

> the generated API documentation covers all functions/classes/etc.

I also included API documentation for the main files cosmos.cpp and cosmos_s.cpp. All other functions are included in the header files. Therefore, the API documentation covers all classes and functions.

> the API documentation and the Wiki pages are linked to from the README.md

I made the links in README.md

> the requirement about having adequate [tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) ideally means that the user can easily verify that the software works correctly, without too much domain knowledge. Right now this is a little difficult: the user has to reproduce figures and it's unclear how much deviation from the results of the figures are acceptable. Can you come up with tests that cover all functionality of the code and the expected behavior is asserted (e.g. code gives you a message/logs about whether it ran correctly)?

I included the code that gives you the differences from the expected results attached as the file "exp_xxx.dat" in all examples. They are automatically generated when the examples are executed. If those values are sufficiently small, like 10^{-10}, the software works correctly. I mentioned it in Wiki.
