### 2025-05-12 #7838 @fruzsinaagocs
Thank you @cmyoo, the package now satisfies the basic doc/test requirements, so I'll start the review now.
