### 2025-05-12 #7838 @Sbozzolo
@fruzsinaagocs, I still find the documentation fragmented and insufficient. I do not wish to review this package in this state, as it would lead to an endless back and forth. I would encourage you to share with the authors resources to guide them in producing useful documentation, if you have some.

In a nutshell, the question I ask myself is "would a technically versed reader be able to do something intersting with this code without consulting the authors and without extensive digging in the source code?" At this point, the answer is still no.

As a new user, I try to follow the wiki. The wiki only gives me three examples and nothing else. The three examples don't have much pedagoical value. they don't teach me how to do much besides running the examples themselves. On the other hand, the API website is just a list of methods that provides no context or usage directions. In addition to this, there's various other documents and website scattered around with various information, including PDFs that compare the existing code with an "original code".

When looking at the files in one of the examples, I see several thousands of lines, various options and cases, blocks commented out. It's unclear how I can extend the examples starting from the source code.

As a user, I don't even know what I can do with COSMOS. Even basic questions are not answered. What configurations can I simulate? What output is produced? How do I change initial conditions/resolution/duration/output variables?

[Here's an example of a numerical relativity code that puts readers in the position to do something interesting with their code without having to reach out to the experts:
https://github.com/GRTLCollaboration/GRChombo]
