### 2025-05-19 #7838 @warrickball
Hi @cmyoo, I thought I'd step in as AEiC to help reach a consensus on what this submissions need to make the review worthwhile. I'll note that we're not creating more work here, as these are issues that would have to be resolved during the course of the review anyway. We're only seeing what we can line up before we initiate the review so that it doesn't quickly lose momentum. Please note the comments already made, which are all valid. In particular, you should certain have another look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and bear in mind that the "OK" requirements should be read as the bare minimum we expect for JOSS.

I'll make the following recommendations and invite @fruzsinaagocs, @Sbozzolo and @ShaunFell to comment, in a bid to reach some consensus on what's needed.

1. It's quite confusing to try to find documentation spread across the [Google site](https://sites.google.com/view/cosmoscode/), [GitHub wiki](https://github.com/cmyoo/cosmos/wiki), [API docs](https://cmyoo.github.io/cosmos/) and PDFs in the example folders. You should **consolidate the main docs to a single location**, with some smaller documentation where necessary to redirect users to those docs. (E.g. from the GitHub README.md to a github.io page). It's fine for these to be in a Wiki, too. (If it's useful, note that a GitHub wiki is itself a git repository; try `git clone https://github.com/cmyoo/cosmos.wiki`)

2. The docs really need something like a **"Getting started" tutorial** that demonstrates how to proceed from a fresh checkout to a working use case. This would address a lack of documentation on [example usage](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage). It's fine to have examples to do this but the missing step is documenting how one would build up the example from scratch. (Perhaps have a look at the  ["Creating your own simulation"](https://github.com/ShaunFell/GRBoondi/wiki/Creating-your-own-simulation) docs from GRBondi as inspiration.)

3. The code layout is itself unclear. E.g. there are source files called `cosmos.cpp` in each example as well as in the `source` folder. It's not clear what is the relationship between them. The docs state a list of files but not their purpose, like what would be required for what sort of calculation. This is also prone to all the `cosmos.cpp` files falling out of sync. Somewhere in the docs (and it might be in a "Getting started" tutorial) we need **a description of the source files** that users will have to interact with. This is true for the parameter files too, for which I didn't easy see any documentation.

As an aside, and knowing that C++ isn't a language I know well, I'd suggest running some sort of code linter/formatter. If users are going to have to interact with the source, it's best that it has a standard style. I don't personally think this a pre-requisite for a review but I'm happy to be overruled, since I'm not a C++ developer.
