### 2025-05-20 #7838 @warrickball
> You don't need to build a huge test suite like GRChombo, but I would recommend following its lead and writing tests for the basic functionality of COSMOS(-S)

While in principle I fully agree, I just want to flag that JOSS doesn't technically require a significant suite of unit tests. Our [review criteria for tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) have as a minimum

> **OK**: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)

Running the examples *might* meet JOSS's minimum requirements but it isn't clear from a glance at the code that the tests will actually fail (e.g. return a non-zero exit code) if the output differs from the reference values. I see what I think are calls to `std::abort` if some I/O fails, but not if the result differs (with some tolerance, presumably).
