### 2025-05-21 #7838 @cmyoo
@warrickball
Thank you for your kind support and suggestions.

As you may already know, we are not software development experts, and our resources are very limited. We understand all the comments so far, but we are discussing with our co-authors what we can and cannot do in the current situation, where the effort we can devote to this development is limited.

We have not yet reached a clear conclusion, but first of all, our code is not designed to provide a complete tool like GRChombo. We expect users to use it by rewriting the source code after understanding it to a certain extent. Therefore, we do not expect that new scientific calculations will be possible just by changing the parameters.

On the other hand, I think that the improvements you proposed are very reasonable. In particular, the second point

> documenting how one would build up the example from scratch. (Perhaps have a look at the ["Creating your own simulation"](https://github.com/ShaunFell/GRBoondi/wiki/Creating-your-own-simulation) docs from GRBondi as inspiration.)

is a perspective that we did not have and is very useful, so I think we can make efforts in this regard by referring to GRBondi. We can also make some efforts to improve the documentation following the referee's suggestions.

Our concern is that we may not be able to meet some of the requests from the referees and the referee candidate with our limited resources. In particular, the referee candidate's requirement, "would a technically versed reader be able to do something intersting with this code without consulting the authors and without extensive digging in the source code" is far removed from our concept of our numerical computation code. Although we understand it might be an important perspective for software development, I think it would be difficult for us, who are not software engineers, to meet this request.
