+++ docs/README.md
@@ -1,0 +1,27 @@
+### Generating documentation
+
+1. Install `sphinx`: 
+```bash
+pip install sphinx
+```
+or install the `requirements.txt`
+
+2. Run `sphinx` quickstart
+```bash
+mkdir docs
+cd docs/
+sphinx-quickstart
+```
+Select `n` for "keeping things together", i.e. no separate source and build directories.
+
+3. Modify the `conf.py` generated by `sphinx`
+According to [this](https://exhale.readthedocs.io/en/latest/quickstart.html) quickstart guide, or based on `oscode`.
+The most important thing here is setting the path to the C++ source and header filed correctly, so that Doxygen can crawl them.
+
+4. Write the `index.rst`
+Add a hyperlink/reference to the api reference documentation, which will be automatically generated by exhale in `api/cosmos-api-reference.rst`.
+
+5. Generate offline documentation
+```bash 
+make html
+```
