+++ docs/README.md
@@ -1,27 +1,36 @@
-### Generating documentation
+# Getting started with automatic documentation using `exhale`
 
-1. Install `sphinx`: 
+This is a skeleton of what the `docs/` directory should look like if you want your C++ docstrings to be automatically formatted by `Doxygen`, and generate a linked set of `html`s. You can then host them on readthedocs, or GitHub pages if you like. Here's how I built this directory.
+
+1. Install `sphinx`, `exhale`, and `sphinx-rtd-theme` (the latter is a stylistic choice, it sets the theme of the pages): 
+
 ```bash
-pip install sphinx
+pip install sphinx exhale sphinx-rtd-theme
 ```
-or install the `requirements.txt`
+or simply
+```bash
+pip install -r requirements.txt`
+```
 
 2. Run `sphinx` quickstart
+
 ```bash
 mkdir docs
 cd docs/
 sphinx-quickstart
 ```
-Select `n` for "keeping things together", i.e. no separate source and build directories.
+In the quickstart, select `n` for "keeping things together", i.e. no separate source and build directories.
 
 3. Modify the `conf.py` generated by `sphinx`
+
 According to [this](https://exhale.readthedocs.io/en/latest/quickstart.html) quickstart guide, or based on `oscode`.
-The most important thing here is setting the path to the C++ source and header filed correctly, so that Doxygen can crawl them.
+The most important thing here is setting the path to the C++ source and header filed correctly, so that `Doxygen` can crawl them.
 
-4. Write the `index.rst`
-Add a hyperlink/reference to the api reference documentation, which will be automatically generated by exhale in `api/cosmos-api-reference.rst`.
+4. Write the `index.rst`, your landing page
+   
+Add a link to the api reference documentation, which will be automatically generated by exhale in `api/cosmos-api-reference.rst`.
 
-5. Generate offline documentation
+6. Generate offline documentation
 ```bash 
 make html
 ```
