+++ docs/README.md
@@ -4,33 +4,39 @@ This is a skeleton of what the `docs/` directory should look like if you want yo
 
 1. Install `sphinx`, `exhale`, and `sphinx-rtd-theme` (the latter is a stylistic choice, it sets the theme of the pages): 
 
-```bash
-pip install sphinx exhale sphinx-rtd-theme
-```
-or simply
-```bash
-pip install -r requirements.txt`
-```
+   ```bash
+   pip install sphinx exhale sphinx-rtd-theme
+   ```
+   or simply
+   ```bash
+   pip install -r requirements.txt`
+   ```
 
 2. Run `sphinx` quickstart
 
-```bash
-mkdir docs
-cd docs/
-sphinx-quickstart
-```
-In the quickstart, select `n` for "keeping things together", i.e. no separate source and build directories.
+   ```bash
+   mkdir docs
+   cd docs/
+   sphinx-quickstart
+   ```
+   In the quickstart, select `n` for "keeping things together", i.e. no separate source and build directories.
 
 3. Modify the `conf.py` generated by `sphinx`
 
-According to [this](https://exhale.readthedocs.io/en/latest/quickstart.html) quickstart guide, or based on `oscode`.
-The most important thing here is setting the path to the C++ source and header filed correctly, so that `Doxygen` can crawl them.
+   According to [this](https://exhale.readthedocs.io/en/latest/quickstart.html) quickstart guide, or based on `oscode`.
+   The most important thing here is setting the path to the C++ source and header filed correctly, so that `Doxygen` can crawl them.
 
 4. Write the `index.rst`, your landing page
    
-Add a link to the api reference documentation, which will be automatically generated by exhale in `api/cosmos-api-reference.rst`.
+   Add a link to the api reference documentation, which will be automatically generated by exhale in `api/cosmos-api-reference.rst`.
 
 6. Generate offline documentation
-```bash 
-make html
-```
+   ```bash 
+   make html
+   ```
+
+To view the docs, navigate to the `_build/html` directory and open `index.html` in your browser. You'll be able to view the interlinked documentation pages here. Specifically, your API documentation is found at "API Reference" under "Contents".
+
+7. Improve the documentation
+
+   `Doxygen` requires [some syntax rules](https://www.doxygen.nl/manual/docblocks.html) to be followed in your classes', functions', etc. docstrings to be able to render the documentation correctly. 
