### 2025-06-01 7ff3000
generation script are added :: makecosmos.sh
