+++ source/scalar_iso/cosmos_initial_scalar.cpp
@@ -1,305 +1,0 @@
-/*********************************************************************************************************/
-/*-------------------------------------------------------------------------------------------------------*/
-/* INITIAL DATA SETTING :: BSSN evolution Class of COSMOS                                                */
-/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
-/*-------------------------------------------------------------------------------------------------------*/
-/*********************************************************************************************************/
-#include "../cosmos.h"
-#include <iomanip>
-#include <cmath>
-#include <fstream>
-#include <cstdio>
-
-using namespace std;
-
-
-void Fmv::initial(double mu,double kk)
-{
-	//setting flatdet and Gamma
-	set_flat();
-	double L=1.;						// box size
-	double inr=0.8;						// Psi=1 for r>inr
-	
-	#pragma omp parallel for
-	for(int l=lli;l<=lui;l++)
-	{
-		double zz=get_z(l);
-		double Z=funcf(zz);
-		
-		for(int k=kli;k<=kui;k++)
-		{
-			double yy=get_y(k);
-			double Y=funcf(yy);
-			
-			for(int j=jli;j<=jui;j++)
-			{
-				double xx=get_x(j);
-				double X=funcf(xx);
-				
-				double r=sqrt(X*X+Y*Y+Z*Z);
-				
-				//see arXiv:2112.12335 for the functional form
-				//we use psi for temporary storage for phi
-				if(r>L)
-				set_psi(l,k,j)=0. ;
-				else if(r<inr)
-				set_psi(l,k,j)=mu*pow(M_E,-(pow(kk,2)*pow(r,2))/6.);
-				else 
-				set_psi(l,k,j)=mu*pow(M_E,-(pow(kk,2)*pow(r,2))/6.)*(1 - pow(inr - L,-36)*pow(pow(inr - L,6) - pow(L - r,6),6));
-			}
-		}
-	}
-	#pragma omp barrier
-	
-	boundary_psi_initial();
-
-	double fGam=1.+fluidw;
-	double Hbi2;
-	double Hbi3;
-
-	cout.precision(8);
-	#pragma omp barrier
-	
-	double fac=2./(3.*fGam);
-
-	double tt=fac/Hb;
-	double HH=fac/tt;
-	double aa=pow(tt/tini,fac);
-	set_t(tt);
-
-	// cout << "tt=" << tt 
-	// << " dt0=" << dt0 
-	// << " HH=" << HH 
-	// << " aa=" << aa 
-	// << " tini=" << tini 
-	// << endl;
-
-	// exit(0);
-
-	Hbi2=pow(HH*aa,-2);
-	Hbi3=pow(HH*aa,-3);
-		
-	#pragma omp parallel for
-	for(int l=lli;l<=lui;l++)
-	{
-		double zz=get_z(l);
-		
[…]
