+++ source/sample_pert/cosmos_bssn.cpp
@@ -1,3675 +1,0 @@
-/*********************************************************************************************************/
-/*-------------------------------------------------------------------------------------------------------*/
-/* BSSN EVOLUTION :: BSSN evolution Class of COSMOS                                                      */
-/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
-/*-------------------------------------------------------------------------------------------------------*/
-/*********************************************************************************************************/
-#include "../cosmos.h"
-//#include "omp.h"
-#include <algorithm>
-
-void Fmv0::BSSN(int itype)
-{
-	//time deviation and fraction for Rungekutta start
-	double frac;
-	// double tt;
-	
-	switch(itype){
-	case 1:
-		dt=0.5*dt0;
-		frac=1./6.;
-		// tt=get_t();
-		break;
-	case 2:
-		dt=0.5*dt0;
-		frac=1./3.;
-		// tt=get_t()+dt;
-		break;
-	case 3:
-		dt=dt0;
-		frac=1./3.;
-		// tt=get_t()+0.5*dt;
-		break;
-	case 4:
-		dt=dt0;
-		frac=1./6.;
-		// tt=get_t()+dt;
-		break;
-	default:
-		frac=1./6.;
-		// tt=get_t();
-	}
-	//time deviation and fraction for Rungekutta end
-		
-	// @@@@@@@@@@@@   variables (BSSN +GAUGE +fluid +scalar)   @@@@@@@@@@@
-	//  0:alpha , 1:bx  ,  2:by  ,  3:bz   ,  -> Gauge
-	//  4:bbx ,  5:bby  ,  6:bbz , -> for hyper-Gamma driver
-	//  7:gxx   , 8:gyy ,  9:gzz , 10:gxy  , 11:gxz , 12:gyz  ,  -> tilde gamma
-	//  13:wa , -> psi 
-	//  14:akxx  ,15:akyy, 16:akzz, 17:akxy , 18:akxz, 19:akyz , -> tilde A_ij
-	//  20:ek  , -> trK
-	// 21:zgx   ,22:zgy , 23:zgz -> tilde Gamma
-	// 24:E, 25:S_x, 26:S_y, 27:S_z, 28:N_B(D), 
-	// 29(25):phi, 30(25) :Pi -> scalar, if no fluid, the numbers are 24 and 25
-	
-	set_zero_d();	//bvd=0
-	
-	//preparation associated with \beta^i
-	#pragma omp barrier
-	BSSN_adv();
-
-	//setting for fluid flux
-	#pragma omp barrier
-	if(fluidevo)
-	{
-		//getting primitive variables from dynamical variables
[…]

[…]
