### 2025-06-01 5d7e425
generation script are added :: makecosmos.sh
