+++ sample/adiabatic_spherical/cosmos.cpp
@@ -1,0 +1,1562 @@
+//**********************************************************************************************//
+//----------------------------------------------------------------------------------------------//
+//                     @@@@@   @@@@    @@@@    @@@ @@    @@@@    @@@@                           //
+//                   @@@      @@   @  @@   @  @@@ @@ @  @@   @  @@   @                          //
+//                   @@       @@   @   @@@    @@  @  @  @@   @   @@@                            //
+//                   @@       @@   @ @    @@  @@     @  @@   @ @    @@                          //
+//                     @@@@@   @@@@   @@@@@   @@     @   @@@@   @@@@@                           //
+//----------------------------------------------------------------------------------------------//
+//**********************************************************************************************//
+//----------------------------------------------------------------------------------------------//
+//                 cosmos: quarter region, mesh-refinement                                      //
+//                                             ver. 1.00           coded by Chulmoon Yoo        //
+//                                                                                              //
+//----------------------------------------------------------------------------------------------//
+// Main file   :: cosmos.cpp                                                                    //
+// header      :: cosmos.h ahf2d.h                                                              //
+// definition  :: cosmos_bssn.cpp cosmos_initial.cpp cosmos_output.cpp                          //
+//                cosmos_boundary.cpp cosmos_ahf.cpp cosmos_ipol.cpp                            //
+//                cosmos_fluid.cpp cosmos_fmr.cpp                                               //
+// how to make :: makefile                                                                      //
+//----------------------------------------------------------------------------------------------//
+//    Compile :: $ make                                                                         //
+//               Choose compiler by changing makefile.                                          //
+//    OpenMP  :: $ export OMP_NUM_THREADS=#                                                     //
+//               Set the number of CPU cores you use.                                           //
+//    Execute :: $ taskset -c 0-# ./cosmos                                                      //
+//               Set CPU cores directly on the machine.                                         //
+//    Clean   :: $ make clean                                                                   //
+//               Delete .o files.                                                               //
+//**********************************************************************************************//
+#include <iostream>
+#include <iomanip>
+#include <fstream>
+#include <sstream>
+#include <cmath>
+#include <cfloat>
+#include <cstring>
+#include "./cosmos.h"			//header for Einstein solver
+#include "./ahf2d.h"			//header for apparent horizon finder
+
+using namespace std;
+
+void check_xkl(Fmv *fmv);
+
[…]
