+++ sample/sample_pert/cosmos_bssn.cpp
@@ -4,7 +4,7 @@
 /*                                             ver. 1.00           coded by Chulmoon Yoo                 */
 /*-------------------------------------------------------------------------------------------------------*/
 /*********************************************************************************************************/
-#include "./cosmos.h"
+#include "cosmos.h"
 //#include "omp.h"
 #include <algorithm>
 
