### 2025-06-01 77e4bf2
modified :: makecosmos.sh
