+++ makecosmos.sh
@@ -10,12 +10,21 @@ if [ -d ${DIR} ]; then
 echo "Directory '${DIR}' already exists.  Please rename it."
 else
 mkdir -p ${DIR}
+
+if [ -d ${DIR} ]; then
+echo "Directory '${DIR}' is created successfully."
+else
+echo "Directory creation error."
+exit
+fi
+
 cp source/makefile ${DIR}/
 cp source/*.h ${DIR}/
 cp source/*.cpp ${DIR}/
 cp sample/${DIR}/*.cpp ${DIR}/
 cp sample/${DIR}/*.d ${DIR}/
 cp sample/${DIR}/*.dat ${DIR}/
+
 fi
 
 elif [ ${DIR} = "sample_pert" ]; then
@@ -24,6 +33,14 @@ if [ -d ${DIR} ]; then
 echo "Directory '${DIR}' already exists.  Please rename it."
 else
 mkdir -p ${DIR}
+
+if [ -d ${DIR} ]; then
+echo "Directory '${DIR}' is created successfully."
+else
+echo "Directory creation error."
+exit
+fi
+
 cp source/makefile ${DIR}/
 cp source/*.h ${DIR}/
 cp source/*.cpp ${DIR}/
@@ -38,6 +55,14 @@ if [ -d ${DIR} ]; then
 echo "Directory '${DIR}' already exists.  Please rename it."
 else
 mkdir -p ${DIR}
+
+if [ -d ${DIR} ]; then
+echo "Directory '${DIR}' is created successfully."
+else
+echo "Directory creation error."
+exit
+fi
+
 cp source/makefile ${DIR}/
 cp source/*.h ${DIR}/
 cp source/*.cpp ${DIR}/
@@ -48,13 +73,23 @@ fi
 
 else
 
-echo "Since there is no sample corresponding to '${DIR}', this script tries to gather necessary files."
-echo "After this step, appropriate initial data should be provided to start your simulation by COSMOS code."
-
 if [ -d ${DIR} ]; then
 echo "Directory '${DIR}' already exists.  Please rename it."
 else
 mkdir -p ${DIR}
+
+if [ -d ${DIR} ]; then
+echo "Directory '${DIR}' is created successfully."
+else
+echo "Directory creation error."
+exit
+fi
+
+echo " "
+echo " "
+echo "Since there is no sample corresponding to '${DIR}', this script tries to gather necessary files."
+echo "After this step, appropriate initial data should be provided to start your simulation by COSMOS code."
+
 cp source/makefile ${DIR}/
 cp source/*.h ${DIR}/
 cp source/*.cpp ${DIR}/
