### 2025-06-08 #7838 @fruzsinaagocs
@cmyoo,
We understand that your time and resources as researchers are limited and in no way expect you to have the expertise of a software engineer. To help reduce your workload, we collected what you need to get a review started and some guidance on how to do it:

* Tests need to raise errors if something is wrong.
    * The tests already compute a difference, so we're looking for something like assert(<condition that evaluates to True if test passes>) , see the [assert documentation](https://en.cppreference.com/w/cpp/error/assert).
* The docs must be consolidated in a single place. Of the options, either a Wiki or readthedocs (or GH pages) seems best.
    * We opened a [pull request](https://github.com/cmyoo/cosmos/pull/1)  for generating API-level documentation with the packages Doxygen + breathe+ exhale, and added a README.md inside the rearranged docs directory that describes how we built the configuration files. You'll need to correct some of the functions'/classes' docstrings so that they are in a Doxygen-readable format. The README.md also contains some links for this. If you merge this pull request, you'll be able to generate HTML pages for your documentation, which you can then host on readthedocs or some other website. We are happy to help set up a readthedocs page (and add automation via GitHub actions). The examples can also be added to the website after being converted to .md or .rst.
    * If you choose to stick with the Wiki, one idea is to generate pages with exhale + Sphinx manually, convert them to Markdown, then pushing to the Wiki repo. This would be easier than syncing by editing Wiki directly.
    * Either way, PDFs in the example folders need to be converted to something else. A README.md or .rst per example folder would be a start, doable with Pandoc if they still have LaTeX source.
* Finally, there needs to be an example of how to get from a clean source directory to a reasonable science application, a "quick start".
    * Consider documenting going from scratch to one of the examples (simplest is fine).

Please let us know if you're willing and able to make these changes, and if so, approximately how much time you'll need.
