+++ sample/flat_simplest_fmr/cosmos.cpp
@@ -36,9 +36,9 @@
 #include <cfloat>
 #include <cstring>
 #include "cosmos.h"			//header for Einstein solver
-// #include "ahf2d.h"			//header for apparent horizon finder
 
 using namespace std;
+
 
 void initial_read(							//reading inital para
 ifstream& fin, 								//initial parameter file
@@ -85,6 +85,30 @@ double& ptintval2,							//2nd
 double& changept							//printing interval change time
 );
 
+void initial_fmr(							//reading initial parameter for fmr
+ifstream& fin,								//initial parameter file for fmr
+int& laymax,								//max fmr layer number
+int *jbs,									//grid number for fmr region on x-axis
+int *kbs,									//grid number for fmr region on y-axis
+int *lbs,									//grid number for fmr region on z-axis
+double *alp_fmr								//values of the lapse for starting fmr
+);
+
+void ahf_read(								//reading initial para for AHF
+ifstream& fin, 								//AHF para file
+double& AHFstart,							// time to start AHF
+int& ahcint, 								//interval step to call AHF
+int& ahpint, 								//interval step to print AH
+int& ntheta, 								//grid num of theta
+int& nphi, 									//grid num of phi
+int& ahloop, 								//max step of AHF
+double& var, 								//parameter \eta_apparent
+double& facn, 								//factor of mix with the next
+double& err_p, 								//error for Poisson solver
+double& err_e, 								//error for AH equation
+double& ahc 								//initial guess of radius
+);
+
 
 int main(int argc,char* argv[])
 {
@@ -93,7 +117,10 @@ int main(int argc,char* argv[])
 		cuev;								//for curvature evaluation
 		
 	long int mstep; 						//max step num for time evo
-
+	// int ahloop, 							//max step of AHF
+		// ahcint, 							//interval step to call AHF
+		// ahpint, 					
+	  //interval step to print AH
 	int	tab, 								//tab num for buf grids
 		nxmin, 								//min grid num of x
 		nxmax, 								//max grid num of x
@@ -101,6 +128,8 @@ int main(int argc,char* argv[])
 		nymax, 								//max grid num of y
 		nzmin, 								//min grid num of z
 	        nzmax; 								//max grid num of z
+		// ntheta, 							//grid num of theta
+		// nphi, 								//grid num of phi
 		// pk,									//printing grid number for y axis
 		// pl;									//printing grid number for z axis
 	
@@ -117,11 +146,17 @@ int main(int argc,char* argv[])
 		zmax, 								//max coord val of z
 		t, 									//time
 		tmax, 								//max time to evolve
+		// var, 								//parameter \eta_apparent
+		// facn, 								//factor of mix with the next
+		// err_p, 								//error for Poisson solver
+		// err_e, 								//error for AH equation
+		// ahc, 								//initial guess of radius
 		cfl, 								//CFL para
 		cdt, 								//factor for cosmological time scale(dt=cfl*cdt*1/H)
[…]
