+++ sample/flat_simplest_fmr/cosmos_fmr.cpp
@@ -1,0 +1,265 @@
+/*********************************************************************************************************/
+/*-------------------------------------------------------------------------------------------------------*/
+/* FUNCTIONS FOR FIXED MESH REFINENMENT :: BSSN evolution Class of COSMOS                                */
+/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
+/*-------------------------------------------------------------------------------------------------------*/
+/*********************************************************************************************************/
+#include "cosmos.h"
+#include <iomanip>
+#include <cmath>
+#include <fstream>
+#include <cstdio>
+
+//time evolution of this layer
+void Fmv1::onestep(int btype)
+{
+	set01();						//xxx1=xxx0 
+	#pragma omp barrier
+	setv0();						//xxx0=xxx 
+	#pragma omp barrier
+	
+	set_zero_r();					////xxxr=0 
+	#pragma omp barrier
+	
+	//each step start
+	BSSN(1);
+	#pragma omp barrier
+	set_boundary(btype,intp);
+	#pragma omp barrier
+	
+	//checking constraint 
+	check_const();
+	#pragma omp barrier
+
+	BSSN(2);
+	#pragma omp barrier
+	set_boundary(btype,intp);
+	#pragma omp barrier
+    BSSN(3);
+	#pragma omp barrier
+    set_boundary(btype+1,intp);
+	#pragma omp barrier
+	BSSN(4);
+	#pragma omp barrier
+	set_boundary(btype+1,intp);
+	#pragma omp barrier
+	//each step end
+
+	set_dtpp(get_dtp());			//one before previous time step
+	set_dtp(get_dt0());				//previous time step
+
+	//upper layers 
+	if(mrf)
+    {
+        ulay->evolve();
+        #pragma omp barrier
+        boundary_quarter();
+    }
+	
+	t+=dt0;							//time forward
+	
+	return;
+}
+
+//refinenment of the lower layer
+void Fmv1::refine_llay()
+{
+	#pragma omp parallel for 
+	for(int lj=ljfmrl;lj<=ljfmru;lj++)
+	{
+		int j=2*lj;
+		for(int ll=llfmrl;ll<=llfmru;ll++)
+		{
+			int l=2*ll;
+			for(int lk=lkfmrl;lk<=lkfmru;lk++)
+			{
+				int k=2*lk;
+				for(int i=0;i<nn;i++)
+				{
+					llay->set_bv(ll,lk,lj,i)=get_bv(l,k,j,i);
+				}
+			}
+		}
+	}
+	
+	return;
+}
+
+//time evolution
+void Fmv1::evolve()
+{
+	dt0=0.5*llay->get_dt0();
+
+    set_boundary(1,-2);
+	#pragma omp barrier
+	onestep(2);
+	#pragma omp barrier
+	onestep(4);
+	#pragma omp barrier
+	refine_llay();
+}
+
+//inital setting when this layer starts
+void Fmv1::set_fmr_initial()
+{	
+	mrf=false;
+
+	llli=llay->get_lli();
+	llui=llay->get_lui();
+	lkli=llay->get_kli();
+	lkui=llay->get_kui();
[…]
