+++ sample/flat_simplest_scaleup/cosmos_output.cpp
@@ -1,0 +1,1401 @@
+/*********************************************************************************************************/
+/*-------------------------------------------------------------------------------------------------------*/
+/* OUTPUT FUNCTIONS :: BSSN evolution Class of COSMOS                                                    */
+/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
+/*-------------------------------------------------------------------------------------------------------*/
+/*********************************************************************************************************/
+#include "cosmos.h"
+#include <iomanip>
+#include <fstream>
+
+//output as functions of x, k:ygrid num, l:z grid num
+void Fmv0::print_x(ofstream& fout, int k, int l)
+{
+	if(l<lmin)
+	{
+		fout << "error : l<lmin" << endl;
+		return;
+
+	}
+	else if(l>lmax)
+	{
+		fout << "error : l>lmax" << endl;
+		return;
+	}
+	else if(k<kmin)
+	{
+		fout << "error : k>kmin" << endl;
+		return;
+	}
+	else if(k>kmax)
+	{
+		fout << "error : k>kmax" << endl;
+		return;
+	}
+
+	double tt = get_t();
+
+	fout.setf(ios_base::fixed, ios_base::floatfield);
+	fout.precision(8);
+	fout << "##" 
+		<< "time=" << tt 
+		<< " trK="  << get_bv(lui,kui,jui,20)
+		<< " k=" << k
+		<< " l=" << l
+		<< endl;
+
+	for(int j=jmin;j<=jmax;j++)
+	{
+		double x = get_x(j);
+		
+		/////////////// output for geometry start ////////////////////////////////////////////////
+		double fxx,fyy,fzz,gxx,gyy,gzz,gxy,gxz,gyz,wa;							//
+		double akxx,akyy,akzz,akxy,akxz,akyz,ek;
+		double zgx,zgy,zgz;
+		double bx,by,bz,bbx,bby,bbz;
+		double M_x,M_y,M_z,nM_x,nM_y,nM_z,dGamx,dGamy,dGamz;
+		double alp,hamn,ham;
+		double Kinv;
+		
+		fxx=get_flat_df2x(j);
+		fyy=get_flat_df2y(k);
+		fzz=get_flat_df2z(l);
+
+		alp= get_bv(l,k,j, 0);
+		bx =  get_bv(l,k,j, 1);
+		by =  get_bv(l,k,j, 2);
+		bz =  get_bv(l,k,j, 3);
+		bbx=  get_bv(l,k,j, 4);
+		bby=  get_bv(l,k,j, 5);
+		bbz=  get_bv(l,k,j, 6);
+		gxx=  fxx+get_bv(l,k,j, 7);
+		gyy=  fyy+get_bv(l,k,j, 8);
+		gzz=  fzz+get_bv(l,k,j, 9);
+		gxy=  get_bv(l,k,j,10);
+		gxz=  get_bv(l,k,j,11);
+		gyz=  get_bv(l,k,j,12);
+		wa=   get_bv(l,k,j,13);
+		akxx= get_bv(l,k,j,14);
+		akyy= get_bv(l,k,j,15);
+		akzz= get_bv(l,k,j,16);
+		akxy= get_bv(l,k,j,17);
+		akxz= get_bv(l,k,j,18);
+		akyz= get_bv(l,k,j,19);
+		ek=   get_bv(l,k,j,20);
+		zgx=  get_bv(l,k,j,21);
[…]

[…]
