### 2025-09-20 09a60bd
[modify] terminal output slightly modified, code-test.yml is modified
