### 2025-09-20 7c39775
[modify] terminal output slightly modified (endl removed)
