### 2025-09-20 eddc9cc
[modify] terminal output slightly modified (endl removed)
