### 2025-09-24 25e6cf9
[Modify] Remove unused functions for boundary conditions
