+++ source/cosmos_boundary.cpp
@@ -1560,254 +1560,6 @@ void Fmv::boundary_periodic()
 	return;
 }
 
-
-//reflection boundary condition for shift
-void Fmv::boundary_beta()
-{
-	short int *wrefsign;
-	
-	// @@@@@@@@@@@@   variables (BSSN +GAUGE)   @@@@@@@@@@@
-	//  0:wx  ,  1:wy  ,  2:wz 
-	
-	int nw=3;
-	
-	wrefsign = new short int[nw];
-	
-	//cout << "boundary reflection" << endl;
-	
-	// x-fixed 
-	/////////////////////////////////////
-	//     0 1 2 3 4 5 6 7 8 9 
-	//     B B B I I I I B B B --> inner and buffer
-	//          |      |       --> boundary position
-	//     2 <= 3, 1 <= 4, 0 <= 5  || jl[j] <= jli+j-1
-	//     7 <= 5, 8 <= 4, 9 <= 3  || ju[j] <= jui-j  for scalar
-	//		
-	//		pure scalar:	0,7,8,9,13,14,15,16,20 
-	//
-	//		vector scalar:	  2,3,    5,6,     22,23
-	//		vector vector:	1,      4,      21,
-	//
-	//		tensor scalar:	      12,        19
-	//		tensor vector:	10,11,     17,18,
-	
-	for(int i=0;i<nw;i++)
-	 wrefsign[i]=1; 
-
-	wrefsign[0]=-1;
-	
-	for(int l=lli;l<=lui;l++)
-	{
-		for(int k=kli;k<=kui;k++)
-		{
-			for(int j=1;j<=tab;j++)
-			{
-				for(int i=0;i<nw;i++)
-				{
-					set_bv(l,k,jl[j],i+1)=wrefsign[i]*get_bv(l,k,jli+j,i+1);
-					set_bv(l,k,ju[j],i+1)=wrefsign[i]*get_bv(l,k,jui-j,i+1);
-				}
-			}
-		}
-	}
-	
-	// y-fixed 
-	//
-	//		pure scalar:	0,7,8,9,13,14,15,16,20 
-	//
-	//		vector scalar:	1,  3,   4,  6,    21,   23
-	//		vector vector:	  2,       5,         22,
-	//
-	//		tensor scalar:	   11,         18,  
-	//		tensor vector:	10,   12,   17,   19
-	for(int i=0;i<nw;i++)
-	 wrefsign[i]=1; 
-
-	wrefsign[1]=-1;
-
-	for(int l=lli;l<=lui;l++)
-	{
-		for(int j=jli;j<=jui;j++)
-		{
-			for(int k=1;k<=tab;k++)
-			{
-				for(int i=0;i<nw;i++)
-				{
-					set_bv(l,kl[k],j,i+1)=wrefsign[i]*get_bv(l,kli+k,-j,i+1);
-					set_bv(l,ku[k],j,i+1)=wrefsign[i]*get_bv(l,kui-k,j,i+1);
-				}
-			}
-		}
-	}
-	
-	// z-fixed
-	//
-	//		pure scalar:	0,7,8,9,13,14,15,16,20 
-	//
-	//		vector scalar:	1,2,    4,5,    21,22,
-	//		vector vector:	    3,      6,        23
-	//
-	//		tensor scalar:	10,         17,  
-	//		tensor vector:	   11,12,      18,19
-	for(int i=0;i<nw;i++)
-	 wrefsign[i]=1; 
-
-	wrefsign[2]=-1;
-
-	for(int k=kli;k<=kui;k++)
-	{
[…]
