### 2025-09-24 9461a63
[edit, remove] paper.md and get_ext functions
