+++ paper/paper.md
@@ -11,10 +11,14 @@ authors:
     orcid: 0000-0002-9928-4757
     # equal-contrib: true
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Hirotada Okawa
+    orcid: 0000-0001-7372-5131
+    # corresponding: true # (This is how to denote the corresponding author)
+    affiliation: 3
   - name: Albert Escrivà
     orcid: 0000-0001-5483-8034
     # equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-    affiliation: 3
+    affiliation: 1
   - name: Tomohiro Harada
     orcid: 0000-0002-9085-9905
     # corresponding: true # (This is how to denote the corresponding author)
@@ -31,10 +35,6 @@ authors:
     orcid: 0000-0002-9579-5787
     # corresponding: true # (This is how to denote the corresponding author)
     affiliation: 6
-  - name: Hirotada Okawa
-    orcid: 0000-0001-7372-5131
-    # corresponding: true # (This is how to denote the corresponding author)
-    affiliation: 7
   - name: Daiki Saito
     orcid: 0000-0003-1624-9268
     # corresponding: true # (This is how to denote the corresponding author)
@@ -56,17 +56,17 @@ affiliations:
    index: 1
  - name: Kobayashi Maskawa Institute, Nagoya University, Japan
    index: 2
-#    ror: 00hx57361
- - name: National Astronomical Observatory of Japan (NAOJ), Japan
+ - name: Faculty of Software and Information Technology, Aomori University, Japan
    index: 3
+#    ror: 00hx57361
+# - name: National Astronomical Observatory of Japan (NAOJ), Japan
+#   index: 3
  - name: Department of Physics, Rikkyo University, Japan
    index: 4
  - name: Center of Gravity, Niels Bohr Institute, Denmark
    index: 5
  - name: Yukawa Institute for Theoretical Physics, Kyoto University, Japan
    index: 6
- - name: Faculty of Software and Information Technology, Aomori University, Japan
-   index: 7
 date: 5 February 2025
 bibliography: paper.bib
 
@@ -84,21 +84,20 @@ or a major component of dark matter.
 In particular, PBHs have been attracting much attention in the recent development of gravitational wave observation. 
 In the standard formation process, PBHs are formed from super-horizon primordial fluctuations with non-linearly large initial amplitude.
 In order to follow the whole non-linear gravitational dynamics, one has to rely on numerical relativity solving Einstein equations.
-`COSMOS` [@Yoo:2013yea; @Okawa:2014nda] and `COSMOS-S` [@Yoo:2021fxs] provide simple tools for the simulation of PBH formation. 
-`COSMOS` and `COSMOS-S` are C++ packages for solving Einstein equations in 3+1 dimensions and spherical symmetry (1+1 dimensions), respectively. 
+`COSMOS` [@Yoo:2013yea; @Okawa:2014nda] provides <!-- and `COSMOS-S` [@Yoo:2021fxs] provide  -->
+ simple tools for the simulation of PBH formation (see [`COSMOS-S`](https://github.com/cmyoo/cosmos-s) for a spherically symmetric version of `COSMOS`). 
+`COSMOS` is a <!-- and `COSMOS-S` are  -->
+C++ packages for solving Einstein equations in 3+1 dimensions. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
[…]
