+++ paper/paper.md
@@ -91,6 +91,7 @@ C++ packages for solving Einstein equations in 3+1 dimensions. <!-- and spherica
 It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specialized for PBH formation. 
 In this paper, we do not describe all scientific results obtained by using `COSMOS`. <!-- or `COSMOS-S`.  -->
 The readers who are interested in the past achievments may refer to @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo; [^1]. <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
+
 [^1]: In most of the references: @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo, additional functions and packages have been implemented to meet the requirements for individual settings. 
 Therefore the results may not be obtained by simply running the public code. 
 
