### 2025-09-24 c0816bf
[Edit] paper.md
