### 2025-09-30 b9f5c22
delete functions without definitions
