### 2025-09-30 #7838 @cmyoo
@fruzsinaagocs,@warrickball
We made major revisions. We decided to separate COSMOS-S and apply only for the publication of COSMOS.
The Wiki for COSMOS has been significantly expanded. The API documentation is now included in the Wiki, and the explanations that were previously separated into PDF files have been moved to the Wiki. In addition, in the "Getting Started" section of the Wiki, after explaining how to perform the simplest test calculations, we have posted tutorials that show examples of how to change the initial conditions and boundary conditions. These examples would work as a guide for users to make use of this code.
This post has not yet reached the review process, but I believe it has undergone at least the minimum revisions required for the review process, so please consider it.
