### 2025-09-30 8ed3dc6
update: cosmos.h
