+++ source/cosmos.h
@@ -2235,6 +2235,9 @@ public:
 	void set_initial_scalar(double mu,double k,double xi2,double xi3);
 	void set_initial_fluid(double mu,double k,double xi2,double xi3);
 	void set_ini_from_Psi();
+        void initial(double mu);
+        void initial(double mu,double kk);
+  
 
 };
 
