### 2025-09-30 613ac5e
modify: makecosmos.sh
