### 2025-12-07 #7838 @fruzsinaagocs
Sorry for the delay, @cmyoo. We think your software now meets the JOSS submission criteria, so the next steps are to close this pre-review and start the review in a different issue. I'll do that now.
