### 2025-12-07 #9570 @editorialbot
[REVIEW]: COSMOS: A numerical relativity code specialized for PBH formation
**Submitting author:** @cmyoo (<a href="http://orcid.org/0000-0002-9928-4757">Chul-Moon Yoo</a>)
**Repository:** https://github.com/cmyoo/cosmos
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
