### 2025-12-08 e9d20c8
Update interpolation functions
