+++ source/cosmos_ipol.cpp
@@ -261,42 +261,37 @@ double facx,double qh,int order,int number)
 	
 
 	////////// shift of grid points and corresponding necessary redefs ///////////////
-	if(fac1<mindis)																	//
+	if(fac1<mindis)
 	{
 		if(eoro==0)
 		{
 			bufdiff++;
-			fac1+=1.;
 			midsta++;
-			fin++;
-			shift++;
 		}
 		else
 		{
-			fac2+=1.;
-			midsta++;
-			fin++;
-			shift++;
+			midend--;
 		}
+		fac1+=1.;
+		shift++;
+		fin++;
 	}
 	if(fac2<mindis)
 	{
 		if(eoro==0)
 		{
-			fac2+=1.;
 			midsta++;
-			fin++;
-			shift++;
 		}
 		else
 		{
 			bufdiff++;
-			fac1+=1.;
-			midsta++;
+			midsta+=2;
+			midend++;
+		}
 			fin++;
 			shift++;
-		}
-	}																				//
+			fac2+=1.;
+	}
 	////////// shift of grid points and corresponding necessary redefs ///////////////
 	
 	
@@ -371,16 +366,11 @@ double facx,double qh,int order,int number)
 	if(order==4)																									//
 	{
 		//4th order finite dif
-		double A=(fac1*(fac2 + pow(fac2,2)))
-				/((1 + fac1)*(2 + 3*fac1 + pow(fac1,2) + 3*fac2 + 2*fac1*fac2 + pow(fac2,2)));
-		double B=-(((1 + fac1)*(fac2 + pow(fac2,2)))
-				/(fac1*(fac1 + pow(fac1,2) + fac2 + 2*fac1*fac2 + pow(fac2,2))));
-		double C=(fac1*(1 + fac1)*(1 + fac2))
-				/(fac2*(fac1 + pow(fac1,2) + fac2 + 2*fac1*fac2 + pow(fac2,2)));
-		double D=-((fac1*(1 + fac1)*fac2)
-				/((1 + fac2)*(2 + 3*fac1 + pow(fac1,2) + 3*fac2 + 2*fac1*fac2 + pow(fac2,2))));
-		double E=(-fac1 - pow(fac1,2) + fac2 - 2*pow(fac1,2)*fac2 + pow(fac2,2) 
-				+ 2*fac1*pow(fac2,2))/(fac1*(1 + fac1)*fac2*(1 + fac2));
+		double A=(fac1*fac2*(1 + fac2))/((1 + fac1)*(1 + fac1 + fac2)*(2 + fac1 + fac2));
+		double B=-(((1 + fac1)*fac2*(1 + fac2))/(fac1*(fac1 + fac2)*(1 + fac1 + fac2)));
+		double C=(fac1*(1 + fac1)*(1 + fac2))/(fac2*(fac1 + fac2)*(1 + fac1 + fac2));
+		double D=-((fac1*(1 + fac1)*fac2)/((1 + fac2)*(1 + fac1 + fac2)*(2 + fac1 + fac2)));
+		double E=-(((fac1 - fac2)*(1 + fac1 + fac2 + 2*fac1*fac2))/(fac1*(1 + fac1)*fac2*(1 + fac2)));
 
 		ans = (A*kzyc[0] +B*kzyc[1] +C*kzyc[2] +D*kzyc[3] +E*qh)*dxi;
 	}
@@ -390,15 +380,14 @@ double facx,double qh,int order,int number)
 		{
 			double A=(fac1*fac2)/((1 + fac1)*(1 + fac1 + fac2));
 			double B=-(((1 + fac1)*fac2)/(fac1*(fac1 + fac2)));
[…]
