+++ README.md
@@ -20,7 +20,7 @@ Detailed installation instructions and usage examples are available in
 our [wiki](https://github.com/GRChombo/GRChombo/wiki), with the home page giving guidance on where to start. -->
 
 ## Documents
-API documents are available in [https://cmyoo.github.io/cosmos/](https://cmyoo.github.io/cosmos/). 
+<!-- API documents are available in [https://cmyoo.github.io/cosmos/](https://cmyoo.github.io/cosmos/).  -->
 
 Wiki pages are available in [https://github.com/cmyoo/cosmos/wiki](https://github.com/cmyoo/cosmos/wiki). 
 
