### 2025-12-08 99dd13e
interpolation update
