+++ source/cosmos.h
@@ -2215,7 +2215,6 @@ public:
 	void boundary_reflection_even(int i);
 	void boundary_prim_reflection();
 	
-	void boundary_reflection_hflags();
 	void boundary_psi_initial();
 	void boundary_reflection_fluid();
 	void boundary_reflection_scalar();
