+++ makecosmos.sh
@@ -17,7 +17,7 @@ echo ${MODEL}
 done
 echo "############################"
 echo "----------------------------"
-echo "If you a new user, please try flat_simplest."
+echo "If you are a new user, please try flat_simplest."
 echo "project name ? > "
 read PNAME
 
