### 2026-02-16 #9570 @fruzsinaagocs
@JesusTorrado yes that would be perfect, thank you!
