### 2026-02-16 #9570 @fruzsinaagocs
👋 @cphyc, would you be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
