### 2026-03-04 cmyoo/cosmos#2 @cphyc
[JOSS-review] Regression testing
I can see there are some automatically-run tests. However, they're just run, and the output isn't checked automatically.
I would encourage you to, at least, read the output and make sure you're getting the same value from one run to another.

Reference:
https://github.com/openjournals/joss-reviews/issues/9570
