### 2026-03-05 #9570 @cmyoo
Hi @cphyc, thank you for your review. I will discuss the issues with my collaborators.
Regarding COSMOS-S, we initially aimed to include it in this submission, but after considering how to publish it in JOSS within our possible efforts, given our limited resources, we decided not to include COSMOS-S in this submission. (Actually, before starting this review process, various points had already been pointed out to us by the potential reviewers, and we have responded to them to the extent possible.) We will try to clarify this point and clearly separate it from COSMOS.
We will also address other issues as soon as possible, so please wait for a while.
