+++ paper/paper.md
@@ -89,14 +89,15 @@ In order to follow the whole non-linear gravitational dynamics, one has to rely 
 `COSMOS` [^1] [@Yoo:2013yea; @Okawa:2014nda] provides <!-- and `COSMOS-S` [@Yoo:2021fxs] provide  -->
  simple tools for the simulation of PBH formation (see `COSMOS-S` [^2] for a spherically symmetric version of `COSMOS`, which we do not deal with in this paper. ). 
 `COSMOS` is a <!-- and `COSMOS-S` are  -->
-C++ packages for solving Einstein equations in 3+1 dimensions. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
+C++ packages for solving Einstein equations in 3+1 dimensions [^3]. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
 It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specialized for PBH formation. 
 In this paper, we do not describe all scientific results obtained by using `COSMOS`. <!-- or `COSMOS-S`.  -->
-The readers who are interested in the past achievements may refer to @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo; [^3]. <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
+The readers who are interested in the past achievements may refer to @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo; [^4]. <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
 
 [^1]: https://github.com/cmyoo/cosmos .
 [^2]: https://github.com/cmyoo/cosmos-s .
-[^3]: In most of the references: @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo, additional functions and packages have been implemented to meet the requirements for individual settings. 
+[^3]: C.Y. and H.O are the main contributors of this code, and other authors used the numerical code during the development and operational stages and contributed in part to its development and improvement.
+[^4]: In most of the references: @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo, additional functions and packages have been implemented to meet the requirements for individual settings. 
 Therefore the results may not be obtained by simply running the public code. 
 
 # Statement of need
@@ -188,21 +189,21 @@ The evolution of sinusoidal small fluctuation is given as an example, which can 
 ### Adiabatic spherically symmetric initial fluctuation
 
 The scalar field is absent in this example. The setting is similar to that in @Yoo:2020lmg.
[…]
