### 2026-03-05 817772c
[edit] paper.md
