### 2026-03-08 c090eb1
check script makefile
