+++ paper/paper.bib
@@ -245,23 +245,31 @@ journal = {Classical and Quantum Gravity}
     pages = "043526",
     year = "2024"
 }
-@article{Escriva:2024lmm,
-    author = "Escriv\`a, Albert and Yoo, Chul-Moon",
-    title = "{Simulations of Ellipsoidal Primordial Black Hole Formation}",
-    eprint = "2410.03452",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
-    month = "10",
-    year = "2024"
-}
 @article{Escriva:2024aeo,
-    author = "Escriv\`a, Albert and Yoo, Chul-Moon",
-    title = "{Non-spherical effects on the mass function of Primordial Black Holes}",
+    author = "Escriv{\`a}, Albert and Yoo, Chul-Moon",
+    title = "{Nonspherical effects on the mass function of primordial black holes}",
     eprint = "2410.03451",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
-    month = "10",
-    year = "2024"
+    doi = "10.1103/4jbp-87wc",
+    journal = "Phys. Rev. D",
+    volume = "112",
+    number = "8",
+    pages = "L081304",
+    year = "2025"
+}
+@article{Escriva:2024lmm,
+    author = "Escriv{\`a}, Albert and Yoo, Chul-Moon",
+    title = "{Simulations of ellipsoidal primordial black hole formation}",
+    eprint = "2410.03452",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1103/PhysRevD.112.083518",
+    journal = "Phys. Rev. D",
+    volume = "112",
+    number = "8",
+    pages = "083518",
+    year = "2025"
 }
 @article{Shimada:2024eec,
     author = "Shimada, Masaaki and Escriv\'a, Albert and Saito, Daiki and Uehara, Koichiro and Yoo, Chul-Moon",
