+++ scalar_iso/cosmos_boundary.cpp
@@ -1,0 +1,2584 @@
+/*********************************************************************************************************/
+/*-------------------------------------------------------------------------------------------------------*/
+/* BOUNDARY CONDITIONS :: BSSN evolution Class of COSMOS                                                 */
+/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
+/*-------------------------------------------------------------------------------------------------------*/
+/*********************************************************************************************************/
+#include "cosmos.h"
+
+//reflection boundary for scalar field
+void Fmv::boundary_reflection_scalar()
+{	
+	#pragma omp parallel for
+	for(int l=lli;l<=lui;l++)
+	{
+		for(int k=kli;k<=kui;k++)
+		{
+			for(int j=1;j<=tab;j++)
+			{
+				for(int i=nsc;i<=nscp;i++)
+				{
+					set_bv(l,k,jl[j],i)=get_bv(l,k,jli+j,i);
+					set_bv(l,k,ju[j],i)=get_bv(l,k,jui-j,i);
+				}
+			}
+		}
+	}
+	
+	// y-fixed 
+	//
+	#pragma omp parallel for
+	for(int j=jli;j<=jui;j++)
+	{
+		for(int l=lli;l<=lui;l++)
+		{
+			for(int k=1;k<=tab;k++)
+			{
+				for(int i=nsc;i<=nscp;i++)
+				{
+					set_bv(l,kl[k],j,i)=get_bv(l,kli+k,-j,i);
+					set_bv(l,ku[k],j,i)=get_bv(l,kui-k,j,i);
+				}
+			}
+		}
+	}
+	
+	// z-fixed
+	//
+	#pragma omp parallel for
+	for(int j=jli;j<=jui;j++)
+	{
+		for(int k=kli;k<=kui;k++)
+		{
+			for(int l=1;l<=tab;l++)
+			{
+				for(int i=nsc;i<=nscp;i++)
+				{
+					set_bv(ll[l],k,j,i)=get_bv(lli+l,k,j,i);
+					set_bv(lu[l],k,j,i)=get_bv(lui-l,k,j,i);
+				}
+			}
+		}
+	}
+	//x,y-fixed
+	
+	#pragma omp parallel for
+	for(int l=lli;l<=lui;l++)
+	{
+		for(int k=1;k<=tab;k++)
+		{
+			for(int j=1;j<=tab;j++)
+			{
+				for(int i=nsc;i<=nscp;i++)
+				{
+					set_bv(l,kl[k],jl[j],i)=get_bv(l,kli+k,-jli-j,i);
+					set_bv(l,ku[k],jl[j],i)=get_bv(l,kui-k,jli+j,i);
+					set_bv(l,kl[k],ju[j],i)=get_bv(l,kli+k,-jui+j,i);
+					set_bv(l,ku[k],ju[j],i)=get_bv(l,kui-k,jui-j,i);
+				}
+			}
+		}
+	}
+
+	// x,z-fixed
+
+	#pragma omp parallel for
+	for(int k=kli;k<=kui;k++)
+	{
+		for(int j=1;j<=tab;j++)
+		{
[…]
