### 2026-03-09 785e0be
[remove] scalar_iso created sample code directory
