+++ scalar_iso/cosmos_initial.cpp
@@ -1,2288 +1,0 @@
-/*********************************************************************************************************/
-/*-------------------------------------------------------------------------------------------------------*/
-/* INITIAL DATA SETTING :: BSSN evolution Class of COSMOS                                                */
-/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
-/*-------------------------------------------------------------------------------------------------------*/
-/*********************************************************************************************************/
-#include "cosmos.h"
-#include <iomanip>
-#include <cmath>
-#include <fstream>
-#include <cstdio>
-
-using namespace std;
-
-//functions for superhorizon initial data with fluid
-//void set_Psi_nonsph(double mu,double k,double xi2,double xi3);
-//void initial_nonsph(double mu,double k,double xi2,double xi3);
-	
-//function to set fluid from geometrical variables by using constraint equations
-//void set_enemomini()
-
-//function to continue the evolution starting from the stored data
-//void initial_continue(char *file_continue);
-
-//functions for inhomogeneous grids
-//double funcf(double X);
-//double ifuncf(double X);
-//double df(double X);
-//double ddf(double X);
-//double dddf(double X);
-//void set_flat();
-
-//function to set initial Gamma^i for evolution (Gamma^i will be treated as dynamical variables)
-//void set_Gam();
-
-//function to set initial scalar field profile
-//void set_initial_scalar(double mu,double k,double xi2,double xi3);
-
-//function to set initial fluid profile
-//void set_initial_fluid(double mu,double k,double xi2,double xi3);
-
-void Fmv::set_Psi_nonsph(double mu,double kk,double xi2,double xi3,double xit2,double xit3,double w)
-{
-	// parameters for the profile: kk, xi2, xi3, xit2, xit3, w(not used)
-	
-	double L=1.;						// box size
-	double inr=0.6;						// Psi=1 for r>inr
-	
-	
-	#pragma omp parallel for
-	for(int l=lli;l<=lui;l++)
-	{
-		double zz=get_z(l);
[…]
