+++ scalar_iso/cosmos_ipol.cpp
@@ -1,1458 +1,0 @@
-/*********************************************************************************************************/
-/*-------------------------------------------------------------------------------------------------------*/
-/* INITERPOLATION FUNCTIONS :: BSSN evolution Class of COSMOS                                            */
-/*                                             ver. 1.00           coded by Chulmoon Yoo                 */
-/*-------------------------------------------------------------------------------------------------------*/
-/*********************************************************************************************************/
-#include "cosmos.h"
-#include <iomanip>
-#include <cmath>
-#include <fstream>
-#include <cstdio>
-
-//interpolation for bv0
-double Fmv0::bv0_ipol(int jc,int kc,int lc,double xc,double yc,double zc,
-				int order,int number)
-//jc,kc,lc:gridnumber of nearest smaller point
-//xc,yc,zc:coordinate value of the point
-//order:order of the interpolation
-//number:specifies the variable for interpolation
-{
-	////////// variables definition and memory allocation ////////////////////////////
-	double ans=0.;										// final answer 
-	double *xx, ***kkk, **kxc, *kxyc;
-	xx  = new double[order];							// coordinate values of grid points
-	kkk  = new double**[order];							// values at points in the cubic region
-	kxc = new double*[order];							// interpolated for x
-	kxyc= new double[order];							// interpolated for x and y
-	int buf=order/2-1;									// number of shift to smaller grids
-	////////// variables definition and memory allocation ////////////////////////////
-
-	////////// preparation of values of variables to be used /////////////////////////
-	//in the interpolation															//
-	for(int l=0;l<order;l++)
-	{
-		kkk[l] = new double*[order];					// memory allocation
-		kxc[l]= new double[order];						// memory allocation
-		xx[l]  =0.;										// initialize
-		kxyc[l]=0.;										// initialize
-		int lt=lc-buf+l;
-		for(int k=0;k<order;k++)
-		{
-			kkk[l][k] = new double[order];				// memory allocation
-			kxc[l][k] = 0.;								// initialize
-			int kt=kc-buf+k;
-			for(int j=0;j<order;j++)
-			{
-				int jt=jc-buf+j;
-				kkk[l][k][j]=get_bv0(lt,kt,jt,number);	// get variables to be used
-			}
-		}
-	}																				//
-	////////// preparation of values of variables to be used /////////////////////////
-
-	////////// x-coordinate interpolation ////////////////////////////////////////////
-	// x-grids																		//
-	for(int j=0;j<order;j++)
-	{
-		xx[j] = get_x(jc-buf+j);
-	}
-	
-	// x-interpolation
-	for(int l=0;l<order;l++)
-	{
-		for(int k=0;k<order;k++)
-		{
-			kxc[l][k] = ipol( xc, xx, kkk[l][k], order );
-		}
-	}																				//
-	////////// x-coordinate interpolation ////////////////////////////////////////////
-
-	////////// y-coordinate interpolation ////////////////////////////////////////////
-	// y-grids																		//
[…]
