+++ .github/workflows/code-test.yml
@@ -2,7 +2,7 @@ name: C/C++ CI
 
 on:
   push:
-    branches: [master, chul]
+    branches: master
 
 jobs:
   adiabatic_spherical:
