### 2026-03-10 cmyoo/cosmos#2 @cmyoo
At the end of the three sample codes (sample_pert, adiabatic_spherical, scalar_iso), I wrote code to read and check the output file. If the output values are the expected values, it will output "Test passed." Accordingly, I have slightly rewritten the relevant part of the wiki.
https://github.com/cmyoo/cosmos/wiki/Evolution-of-a-single-mode-perturbation
https://github.com/cmyoo/cosmos/wiki/Adiabatic-spherically-symmetric-initial-fluctuation
https://github.com/cmyoo/cosmos/wiki/Spherically-symmetric-isocurvature
