### 2026-03-17 #9570 @JesusTorrado
My review follows up on @cphyc's one, understanding that the (manual) testing issue can be considered addressed. I have also added an issue about versioning.
