### 2026-03-23 #9570 @cmyoo
> My review follows up on [@cphyc](https://github.com/cphyc)'s one, understanding that the (manual) testing issue can be considered addressed. I have also added an issue about versioning.

Thank you for your suggestion.
We have released the first version as v1.0.0.
