### 2026-03-24 #9570 @cmyoo
Thank you @JesusTorrado .
I added the release information to the README file.
