### 2026-04-20 #9570 @fruzsinaagocs
Hi @cmyoo, and apologies for the delay in processing your submission. I am currently traveling but looking at the state of the review, all we need is for @cphyc to confirm that they are happy to recommend the submission for acceptance (since all their review points are checked off). Then I will do one final review of the paper and suggest any corrections before asking you to create an archive and release, and then finally hand it over to the editorial team for acceptance.

@cphyc are you happy with the resolution of the issues you raised and to recommend the paper for acceptance?
