### 2026-04-20 #9570 @cmyoo
Hi @cphyc , @fruzsinaagocs ,
Thank you for your assistance in your busy schedule. I have also corrected the citation formatting issue you pointed out, so please check it.
