### 2026-05-13 f402f88
[edit] paper.md
