+++ paper/paper.md
@@ -80,99 +80,95 @@ bibliography: paper.bib
 
 # Summary
 
-Primordial black holes (PBHs) are black holes generated in the early universe without experience of the form of a star.
-It has been pointed out that PBHs may be candidates for black holes and compact objects of various masses in the universe
+Primordial black holes (PBHs) are black holes generated in the early universe without having gone through stellar evolution.
+It has been hypothesized that PBHs may be candidates for black holes and compact objects of various masses in the universe
 or a major component of dark matter. 
 In particular, PBHs have been attracting much attention in the recent development of gravitational wave observation. 
 In the standard formation process, PBHs are formed from super-horizon primordial fluctuations with non-linearly large initial amplitude.
-In order to follow the whole non-linear gravitational dynamics, one has to rely on numerical relativity solving Einstein equations.
+In order to simulate the non-linear gravitational dynamics of PBH formation, one has to rely on numerical relativity solvers to approximate the solution of the Einstein equations.
 `COSMOS` [^1] [@Yoo:2013yea; @Okawa:2014nda] provides <!-- and `COSMOS-S` [@Yoo:2021fxs] provide  -->
- simple tools for the simulation of PBH formation (see `COSMOS-S` [^2] for a spherically symmetric version of `COSMOS`, which we do not deal with in this paper. ). 
+ simple tools for the simulation of PBH formation (see `COSMOS-S` [^2] for a spherically symmetric version of `COSMOS`, which is not discussed in this paper). 
 `COSMOS` is a <!-- and `COSMOS-S` are  -->
-C++ packages for solving Einstein equations in 3+1 dimensions [^3]. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
-It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specialized for PBH formation. 
-In this paper, we do not describe all scientific results obtained by using `COSMOS`. <!-- or `COSMOS-S`.  -->
-The readers who are interested in the past achievements may refer to @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo [^4]. <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
+C++ package for solving the Einstein equations in 3+1 dimensions. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
+It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specifically for the simulation of PBH formation [^3]. 
+Past publications that use COSMOS for simulation include @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo [^4]. <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
 
 [^1]: https://github.com/cmyoo/cosmos .
 [^2]: https://github.com/cmyoo/cosmos-s .
 [^3]: C.Y. and H.O are the main contributors of this code, and other authors used the numerical code during the development and operational stages and contributed in part to its development and improvement.
[…]
