+++ paper/paper.bib
@@ -52,7 +52,7 @@
 }
 @article{Yoo:2018pda,
     author = "Yoo, Chul-Moon and Ikeda, Taishi and Okawa, Hirotada",
-    title = "{Gravitational Collapse of a Massless Scalar Field in a Periodic Box}",
+    title = "{Gravitational collapse of a massless scalar field in a periodic box}",
     eprint = "1811.00762",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
@@ -74,7 +74,7 @@
  }
 @Book{gourgoulhon20123+1,
  author = {Gourgoulhon, Eric},
- title = {3+1 formalism in general relativity : bases of numerical relativity},
+ title = {3+1 Formalism in general relativity : bases of numerical relativity},
  publisher = {Springer},
  year = {2012},
  address = {Berlin New York},
@@ -113,7 +113,7 @@
 }
 @article{Yoo:2020lmg,
     author = "Yoo, Chul-Moon and Harada, Tomohiro and Okawa, Hirotada",
-    title = "{Threshold of Primordial Black Hole Formation in Nonspherical Collapse}",
+    title = "{Threshold of primordial black hole formation in nonspherical collapse}",
     eprint = "2004.01042",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
@@ -220,7 +220,7 @@ journal = {Classical and Quantum Gravity}
 }
 @article{Yoo:2016kzu,
     author = "Yoo, Chul-Moon and Harada, Tomohiro and Okawa, Hirotada",
-    title = "{3D Simulation of Spindle Gravitational Collapse of a Collisionless Particle System}",
+    title = "{3D Simulation of spindle gravitational collapse of a collisionless particle system}",
     eprint = "1611.07906",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
@@ -297,7 +297,7 @@ journal = {Classical and Quantum Gravity}
 }
 @article{Kurganov:2000ovy,
     author = "Kurganov, Alexander and Tadmor, Eitan",
-    title = "{New High-Resolution Central Schemes for Nonlinear Conservation Laws and Convection\textendash{}Diffusion Equations}",
+    title = "{New high-resolution central schemes for nonlinear conservation laws and convection\textendash{}diffusion equations}",
     doi = "10.1006/jcph.2000.6459",
     journal = "J. Comput. Phys.",
     volume = "160",
