+++ paper/paper.md
@@ -186,34 +186,27 @@ The evolution of a small sinusoidal fluctuation is given as an example, which ca
 
 Here, we consider the adiabatic perturbation generated by the initial curvature perturbation without scalar field contribution. 
 As is described in @Harada:2015yda, once the spatial profile of the curvature perturbation is specified,
-the growing mode solution can be described in the long-wavelength approximation.
-We use the analytic expressions for the geometrical variables with the constant-mean-curvature and zero-shift gauge 
-as the initial data for this example. More details including the specific functional form of the curvature perturbation can be found in [^5] (see also @Yoo:2020lmg). 
-
-In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions [^6]). 
+the growing mode solution can be described in the long-wavelength approximation. 
+More details including the specific functional form of the curvature perturbation can be found in [^5] (see also @Yoo:2020lmg). 
+In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions [^5]). 
 
 ![The lapse function (``$tt$-component" of the metric) on the $xy$-plane at the time when an apparent horizon is found. The blue, green, and purple meshes show the region covered by the lowest, 1st, and 2nd mesh refinement layers, respectively.\label{fig:alp}](alp_2d.pdf){height="8cm"}
 
 ![The shape of the apparent horizon when it is found.\label{fig:AH}](AH_tex.pdf){height="6cm"}
 
-[^5]: https://github.com/cmyoo/cosmos/wiki/Note-for-the-time-evolution-sample-of-an-adiabatic-spherical-case .
-[^6]: https://github.com/cmyoo/cosmos/wiki/Adiabatic-spherically-symmetric-initial-fluctuation .
+[^5]: https://github.com/cmyoo/cosmos/wiki/Adiabatic-spherically-symmetric-initial-fluctuation .
 
 ### Spherically symmetric iso-curvature
 
 Here, we consider the iso-curvature perturbation generated by a massless scalar field.
 We assume that the massless scalar field does not contribute to the background metric of the gradient expansion.
-Then, as is described in @Yoo:2021fxs, once the
+Then, as is described in @Yoo:2021fxs, once the 
 spatial profile of the scalar field is specified at the leading order of the gradient expansion,
 the growing mode solution can be described in the long-wavelength approximation.
-We use the analytic expressions for the geometrical variables with the constant-mean-curvature and zero-shift gauge
-as the initial data for this example. More details including the specific functional form of the curvature perturbation can be found in [^7] (see also @Yoo:2021fxs).
-
+More details including the specific functional form of the curvature perturbation can be found in [^6] (see also @Yoo:2021fxs).
 In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found, as for the adiabatic case. 
 
-
-[^7]: https://github.com/cmyoo/cosmos/wiki/Note-for-the-time-evolution-sample-of-a-massless-scalar-iso-curvature-system .
-[^8]: https://github.com/cmyoo/cosmos/wiki/Spherically-symmetric-isocurvature .
+[^6]: https://github.com/cmyoo/cosmos/wiki/Spherically-symmetric-isocurvature .
 
 
 <!--
