+++ paper/paper.md
@@ -188,7 +188,7 @@ Here, we consider the adiabatic perturbation generated by the initial curvature 
 As is described in @Harada:2015yda, once the spatial profile of the curvature perturbation is specified,
 the growing mode solution can be described in the long-wavelength approximation. 
 More details including the specific functional form of the curvature perturbation can be found in [^5] (see also @Yoo:2020lmg). 
-In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions [^5]). 
+In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions in the example code).  
 
 ![The lapse function (``$tt$-component" of the metric) on the $xy$-plane at the time when an apparent horizon is found. The blue, green, and purple meshes show the region covered by the lowest, 1st, and 2nd mesh refinement layers, respectively.\label{fig:alp}](alp_2d.pdf){height="8cm"}
 
@@ -202,7 +202,7 @@ Here, we consider the iso-curvature perturbation generated by a massless scalar 
 We assume that the massless scalar field does not contribute to the background metric of the gradient expansion.
 Then, as is described in @Yoo:2021fxs, once the 
 spatial profile of the scalar field is specified at the leading order of the gradient expansion,
-the growing mode solution can be described in the long-wavelength approximation.
+the growing mode solution can be described in the long-wavelength approximation. 
 More details including the specific functional form of the curvature perturbation can be found in [^6] (see also @Yoo:2021fxs).
 In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found, as for the adiabatic case. 
 
