+++ paper/paper.md
@@ -190,9 +190,9 @@ the growing mode solution can be described in the long-wavelength approximation.
 More details including the specific functional form of the curvature perturbation can be found in [^5] (see also @Yoo:2020lmg). 
 In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions in the example code).  
 
-![The lapse function (``$tt$-component" of the metric) on the $xy$-plane at the time when an apparent horizon is found. The blue, green, and purple meshes show the region covered by the lowest, 1st, and 2nd mesh refinement layers, respectively.\label{fig:alp}](alp_2d.pdf){height="8cm"}
+![The lapse function (``$tt$-component" of the metric) on the $xy$-plane at the time when an apparent horizon is found. The blue, green, and purple meshes show the region covered by the lowest, 1st, and 2nd mesh refinement layers, respectively.\label{fig:alp}](alp_2d.pdf){height="7cm"}
 
-![The shape of the apparent horizon when it is found.\label{fig:AH}](AH_tex.pdf){height="6cm"}
+![The shape of the apparent horizon when it is found.\label{fig:AH}](AH_tex.pdf){height="5cm"}
 
 [^5]: https://github.com/cmyoo/cosmos/wiki/Adiabatic-spherically-symmetric-initial-fluctuation .
 
