+++ paper/paper.md
@@ -167,7 +167,7 @@ This approach differs from the standard method of obtaining the initial data for
 # Examples
 
 Three examples are included in the package.
-These examples are intended primarily for demonstration and instructional purposes, and thus, and thus they are not highly accurate. 
+These examples are intended primarily for demonstration and instructional purposes, and thus they are not highly accurate. 
 The resolution has been intentionally kept as low as possible. 
 In the figures below, the length scale is normalized by the size $L$ of the box for the numerical simulation. 
 
@@ -185,10 +185,10 @@ The evolution of a small sinusoidal fluctuation is given as an example, which ca
 ### Adiabatic spherically symmetric initial fluctuation
 
 Here, we consider the adiabatic perturbation generated by the initial curvature perturbation without scalar field contribution. 
-As is described in @Harada:2015yda, once the spatial profile of the curvature perturbation is specified,
+As is described in @Harada:2015yda, once the spatial profile of the curvature perturbation $\zeta(\vec x)$ is specified,
 the growing mode solution can be described in the long-wavelength approximation. 
-More details including the specific functional form of the curvature perturbation can be found in [^5] (see also @Yoo:2020lmg). 
-In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions in the example code).  
+More details including the specific functional form of the curvature perturbation can be found in the instruction page[^5] (see also @Yoo:2020lmg). 
+In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found (see \autoref{fig:alp} and \autoref{fig:AH} which can be generated by following the instructions).  
 
 ![The lapse function (``$tt$-component" of the metric) on the $xy$-plane at the time when an apparent horizon is found. The blue, green, and purple meshes show the region covered by the lowest, 1st, and 2nd mesh refinement layers, respectively.\label{fig:alp}](alp_2d.pdf){height="7cm"}
 
@@ -201,9 +201,9 @@ In the repository, we include the data file `ini_all.dat` necessary to reconstru
 Here, we consider the iso-curvature perturbation generated by a massless scalar field.
 We assume that the massless scalar field does not contribute to the background metric of the gradient expansion.
 Then, as is described in @Yoo:2021fxs, once the 
-spatial profile of the scalar field is specified at the leading order of the gradient expansion,
+spatial profile of the iso-curvature mode $\Upsilon(\vec x)$ is specified at the leading order of the gradient expansion,
 the growing mode solution can be described in the long-wavelength approximation. 
-More details including the specific functional form of the curvature perturbation can be found in [^6] (see also @Yoo:2021fxs).
+More details including the specific functional form of $\Upsilon(\vec x)$ can be found in the instruction page[^6] (see also @Yoo:2021fxs).
 In the repository, we include the data file `ini_all.dat` necessary to reconstruct the geometry and matter distribution at the time an apparent horizon is found, as for the adiabatic case. 
 
 [^6]: https://github.com/cmyoo/cosmos/wiki/Spherically-symmetric-isocurvature .
