+++ paper/paper.bib
@@ -232,19 +232,6 @@ journal = {Classical and Quantum Gravity}
     pages = "105010",
     year = "2017"
 }
-@article{Yoo:2024lhp,
-    author = "Yoo, Chul-Moon",
-    title = "{Primordial black hole formation from a nonspherical density profile with a misaligned deformation tensor}",
-    eprint = "2403.11147",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
-    doi = "10.1103/PhysRevD.110.043526",
-    journal = "Phys. Rev. D",
-    volume = "110",
-    number = "4",
-    pages = "043526",
-    year = "2024"
-}
 @article{Escriva:2024aeo,
     author = "Escriv{\`a}, Albert and Yoo, Chul-Moon",
     title = "{Nonspherical effects on the mass function of primordial black holes}",
