### 2026-05-29 2cfe398
Initial editorial tweaks for publication in JOSS
