+++ paper/paper.bib
@@ -5,19 +5,19 @@
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     doi = "10.1103/PhysRevD.78.064054",
-    journal = "Phys. Rev. D",
+    journal = "Physical Review D",
     volume = "78",
     pages = "064054",
     year = "2008"
 }
 @article{Okawa:2014nda,
     author = "Okawa, Hirotada and Witek, Helvi and Cardoso, Vitor",
-    title = "{Black holes and fundamental fields in Numerical Relativity: initial data construction and evolution of bound states}",
+    title = "{Black holes and fundamental fields in numerical relativity: initial data construction and evolution of bound states}",
     eprint = "1401.1548",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     doi = "10.1103/PhysRevD.89.104032",
-    journal = "Phys. Rev. D",
+    journal = "Physical Review D",
     volume = "89",
     number = "10",
     pages = "104032",
@@ -25,13 +25,13 @@
 }
 @article{Yoo:2013yea,
     author = "Yoo, Chul-Moon and Okawa, Hirotada and Nakao, Ken-ichi",
-    title = "{Black Hole Universe: Time Evolution}",
+    title = {Black Hole Universe: Time Evolution},
     eprint = "1306.1389",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     reportNumber = "OCU-PHYS:382, AP-GR:105, YITP-13-45",
     doi = "10.1103/PhysRevLett.111.161102",
-    journal = "Phys. Rev. Lett.",
+    journal = "Physical Review Letters",
     volume = "111",
     pages = "161102",
     year = "2013"
@@ -44,7 +44,7 @@
     primaryClass = "gr-qc",
     reportNumber = "YITP-21-161, RUP-21-23",
     doi = "10.1103/PhysRevD.105.103538",
-    journal = "Phys. Rev. D",
+    journal = "Physical Review D",
     volume = "105",
     number = "10",
     pages = "103538",
@@ -57,7 +57,7 @@
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     doi = "10.1088/1361-6382/ab06e2",
-    journal = "Class. Quant. Grav.",
+    journal = "Classical and Quantum Gravity",
     volume = "36",
     number = "7",
     pages = "075004",
@@ -88,7 +88,7 @@
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     doi = "10.1103/PhysRevD.110.043526",
-    journal = "Phys. Rev. D",
+    journal = "Physical Review D",
     volume = "110",
     number = "4",
     pages = "043526",
@@ -99,7 +99,7 @@
                         and Koga, Yasutaka",
       title          = "{Cosmological long-wavelength solutions and primordial
                         black hole formation}",
-      journal        = "Phys. Rev.",
+      journal        = "Physical Review D",
       volume         = "D91",
       year           = "2015",
       number         = "8",
@@ -119,7 +119,7 @@
     primaryClass = "gr-qc",
     reportNumber = "RUP-20-12",
     doi = "10.1103/PhysRevD.102.043526",
-    journal = "Phys. Rev. D",
+    journal = "Physical Review D",
     volume = "102",
     number = "4",
     pages = "043526",
@@ -134,7 +134,7 @@
     primaryClass = "gr-qc",
     reportNumber = "RUP-24-1",
     doi = "10.1088/1475-7516/2025/01/003",
-    journal = "JCAP",
+    journal = "Journal of Cosmology and Astroparticle Physics",
     volume = "01",
[…]
