+++ paper/paper.md
@@ -65,9 +65,9 @@ affiliations:
    index: 4
  - name: Center of Gravity, Niels Bohr Institute, Denmark
    index: 5
- - name: Department of Physics, College of Humanities　and Sciences, Nihon University, Japan
+ - name: Department of Physics, College of Humanities and Sciences, Nihon University, Japan
    index: 6
- - name: Department of Science Education, Ewha Womans University, Korea
+ - name: Department of Science Education, Ewha Womans University, Republic of Korea
    index: 7
 date: 5 February 2025
 bibliography: paper.bib
@@ -86,22 +86,20 @@ or a major component of dark matter.
 In particular, PBHs have been attracting much attention in the recent development of gravitational wave observation. 
 In the standard formation process, PBHs are formed from super-horizon primordial fluctuations with non-linearly large initial amplitude.
 In order to simulate the non-linear gravitational dynamics of PBH formation, one has to rely on numerical relativity solvers to approximate the solution of the Einstein equations.
-`COSMOS` [^1] [@Yoo:2013yea; @Okawa:2014nda] provides <!-- and `COSMOS-S` [@Yoo:2021fxs] provide  -->
- simple tools for the simulation of PBH formation (see `COSMOS-S` [^2] for a spherically symmetric version of `COSMOS`, which is not discussed in this paper). 
-`COSMOS` is a <!-- and `COSMOS-S` are  -->
+[COSMOS](https://github.com/cmyoo/cosmos) [@Yoo:2013yea; @Okawa:2014nda] provides <!-- and COSMOS-S [@Yoo:2021fxs] provide  -->
+ simple tools for the simulation of PBH formation (see [COSMOS-S](https://github.com/cmyoo/cosmos-s) for a spherically symmetric version of COSMOS, which is not discussed in this paper).
+COSMOS is a <!-- and COSMOS-S are  -->
 C++ package for solving the Einstein equations in 3+1 dimensions. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
-It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specifically for the simulation of PBH formation [^3]. 
-Past publications that use COSMOS for simulation include @Yoo:2013yea; @Okawa:2014nda; @Yoo:2014boa; @Okawa:2014sxa; @Ikeda:2015hqa; @Brito:2015yga; @Brito:2015yfh; @Okawa_2015; @Yoo:2016kzu; @Yoo:2018pda; @Yoo:2024lhp; @Escriva:2024lmm; @Escriva:2024aeo [^4]. <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
+It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specifically for the simulation of PBH formation.[^1]
+Past publications that use COSMOS for simulation include @Yoo:2013yea, @Okawa:2014nda, @Yoo:2014boa, @Okawa:2014sxa, @Ikeda:2015hqa, @Brito:2015yga, @Brito:2015yfh, @Okawa_2015, @Yoo:2016kzu, @Yoo:2018pda, @Yoo:2024lhp, @Escriva:2024lmm and @Escriva:2024aeo.[^2] <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
 
-[^1]: https://github.com/cmyoo/cosmos .
-[^2]: https://github.com/cmyoo/cosmos-s .
-[^3]: C.Y. and H.O are the main contributors of this code, and other authors used the numerical code during the development and operational stages and contributed in part to its development and improvement.
-[^4]: In these works, additional functions and packages have been implemented that may not appear in the public release of COSMOS. Therefore the results may not be obtained by simply running the public code. 
[…]
