+++ paper/paper.md
@@ -86,20 +86,18 @@ or a major component of dark matter.
 In particular, PBHs have been attracting much attention in the recent development of gravitational wave observation. 
 In the standard formation process, PBHs are formed from super-horizon primordial fluctuations with non-linearly large initial amplitude.
 In order to simulate the non-linear gravitational dynamics of PBH formation, one has to rely on numerical relativity solvers to approximate the solution of the Einstein equations.
-[COSMOS](https://github.com/cmyoo/cosmos) [@Yoo:2013yea; @Okawa:2014nda] provides <!-- and COSMOS-S [@Yoo:2021fxs] provide  -->
- simple tools for the simulation of PBH formation (see [COSMOS-S](https://github.com/cmyoo/cosmos-s) for a spherically symmetric version of COSMOS, which is not discussed in this paper).
-COSMOS is a <!-- and COSMOS-S are  -->
-C++ package for solving the Einstein equations in 3+1 dimensions. <!-- and spherical symmetry (1+1 dimensions), respectively.  -->
+[COSMOS](https://github.com/cmyoo/cosmos) [@Yoo:2013yea; @Okawa:2014nda] provides simple tools for the simulation of PBH formation (see [COSMOS-S](https://github.com/cmyoo/cosmos-s) for a spherically symmetric version of COSMOS, which is not discussed in this paper). 
+COSMOS is a C++ package for solving the Einstein equations in 3+1 dimensions. 
 It was originally translated from SACRA code [@Yamamoto:2008js] into C++ and has been developed specifically for the simulation of PBH formation.[^1]
-Past publications that use COSMOS for simulation include @Yoo:2013yea, @Okawa:2014nda, @Yoo:2014boa, @Okawa:2014sxa, @Ikeda:2015hqa, @Brito:2015yga, @Brito:2015yfh, @Okawa_2015, @Yoo:2016kzu, @Yoo:2018pda, @Yoo:2024lhp, @Escriva:2024lmm and @Escriva:2024aeo.[^2] <!-- @Yoo:2021fxs; @Shimada:2024eec  -->
+Past publications that use COSMOS for simulation include @Yoo:2013yea, @Okawa:2014nda, @Yoo:2014boa, @Okawa:2014sxa, @Ikeda:2015hqa, @Brito:2015yga, @Brito:2015yfh, @Okawa_2015, @Yoo:2016kzu, @Yoo:2018pda, @Yoo:2024lhp, @Escriva:2024lmm and @Escriva:2024aeo.[^2] 
 
-[^1]: C.Y. and H.O are the main contributors of this code, and other authors used the numerical code during the development and operational stages and contributed in part to its development and improvement.
-[^2]: In these works, additional functions and packages have been implemented that may not appear in the public release of COSMOS. Therefore the results may not be obtained by simply running the public code.
+[^1]: C.Y. and H.O. are the main contributors of this code, and other authors used the numerical code during the development and operational stages and contributed in part to its development and improvement.
+[^2]: In these works, additional functions and packages have been implemented that may not appear in the public release of COSMOS. Therefore, the results may not be obtained by simply running the public code.
 
 # Statement of need
 
 In the simulation of PBH formation, the presence of multiple length scales (the size of the collapsing region and that of cosmological expansion) necessitates an efficient resolution refinement procedure. 
-In order to resolve the collapsing region, non-Cartesian scale-up coordinates [@Yoo:2018pda] and a fixed mesh-refinement procedure [@Yoo:2024lhp] are implemented in COSMOS. <!-- The 1+1 dimensional simulation code COSMOS-S [@Yoo:2021fxs] is derived from COSMOS with the CARTOON method [@Alcubierre:1999ab].  -->
[…]
