### 2026-05-31 8b6713b
add paper DOI to readme
